% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\docType{methods}
\name{add_filter,tiledb_filter_list,tiledb_filter-method}
\alias{add_filter,tiledb_filter_list,tiledb_filter-method}
\title{Add a filter to a filter list}
\usage{
\S4method{add_filter}{tiledb_filter_list,tiledb_filter}(object, filter)
}
\arguments{
\item{object}{tiledb_filter_list}

\item{filter}{tiledb_filter_list to add}
}
\description{
Add a filter to a filter list
}
\examples{
flt <- tiledb_filter("ZSTD")
filter_list <- tiledb_filter_list(ctx)
filter_list.add_filter(flt)

}
