% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{as.vector.tiledb_config}
\alias{as.vector.tiledb_config}
\title{Convert a `tiledb_config`` object to an R named character vector}
\usage{
\method{as.vector}{tiledb_config}(x, mode = "any")
}
\arguments{
\item{x}{\code{tiledb_config} object}
}
\value{
a character vector of config parameter names, values
}
\description{
Convert a `tiledb_config`` object to an R named character vector
}
\examples{
cfg <- tiledb_config()
as.vector(cfg)

}
