% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{as_data_table}
\alias{as_data_table}
\title{Construct a data.table from variable-length array query results}
\usage{
as_data_table(dom, data)
}
\arguments{
\item{dom}{tiledb_domain object}

\item{data}{tiledb object to be converted}
}
\value{
data.frame object constructed from \code{data}
}
\description{
Converts a tiledb object to a data.table object suitable for handling
variable-length arrays
}
