% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_uniform_dimensions}
\alias{assert_uniform_dimensions}
\title{Check for unifomity of array dims for a list of arrays}
\usage{
assert_uniform_dimensions(x)
}
\arguments{
\item{x}{a named list of arrays or vectors}
}
\value{
TRUE (or error)
}
\description{
Given a list of arrays, make sure they all have the same dimensions,
and thus are viable input for creating or updating a multi-attribute
TileDB array.
}
\examples{
assert_uniform_dimensions( list(a = matrix(1:4,ncol=2), b = matrix(1:4,ncol=2)))
}
