% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{attrs<-,tiledb_array-method}
\alias{attrs<-,tiledb_array-method}
\title{Selects attributes for the given TileDB array}
\usage{
\S4method{attrs}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A character vector with attributes; the value \code{NA_character_}
signals no attributes should be returned; default is an empty character vector
implying all columns are returned.}
}
\value{
The modified \code{tiledb_array} object
}
\description{
Selects attributes for the given TileDB array
}
