% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{attrs<-}
\alias{attrs<-}
\alias{attrs<-,tiledb_dense-method}
\title{Selects attributes for the given TileDB array}
\usage{
attrs(x) <- value

\S4method{attrs}{tiledb_dense}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_dense} array object}

\item{value}{A character vector with attributes}
}
\value{
The modified \code{tiledb_dense} array object
}
\description{
Selects attributes for the given TileDB array
}
