% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{attrs<-,tiledb_sparse-method}
\alias{attrs<-,tiledb_sparse-method}
\title{Selects attributes for the given TileDB array}
\usage{
\S4method{attrs}{tiledb_sparse}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_sparse} array object}

\item{value}{A character vector with attributes}
}
\value{
The modified \code{tiledb_sparse} array object
}
\description{
Selects attributes for the given TileDB array
}
