% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{attrs,tiledb_array,ANY-method}
\alias{attrs,tiledb_array,ANY-method}
\title{Retrieve attributes from \code{tiledb_array} object}
\usage{
\S4method{attrs}{tiledb_array,ANY}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
An empty character vector if no attributes have been selected or else
a vector with attributes; \code{NA} means no attributes will be returned.
}
\description{
By default, all attributes will be selected. But if a subset of attribute
names is assigned to the internal slot \code{attrs}, then only those attributes
will be queried.  This methods accesses the slot.
}
