% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\docType{methods}
\name{attrs,tiledb_array_schema,ANY-method}
\alias{attrs,tiledb_array_schema,ANY-method}
\title{Returns a list of all \code{tiledb_attr} objects associated with the \code{tiledb_array_schema}}
\usage{
\S4method{attrs}{tiledb_array_schema,ANY}(object, idx, ...)
}
\arguments{
\item{object}{tiledb_array_schema}
}
\value{
a list of tiledb_attr objects
}
\description{
Returns a list of all \code{tiledb_attr} objects associated with the \code{tiledb_array_schema}
}
\examples{
ctx <- tiledb_ctx()
dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(1L, 10L), type = "INT32")))
sch <- tiledb_array_schema(ctx, dom, attrs = c(tiledb_attr(ctx, "a1", type = "INT32"),
                                               tiledb_attr(ctx, "a2", type = "FLOAT64")))
attrs(sch)

lapply(attrs(sch), datatype)

}
