% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{capacity<-}
\alias{capacity<-}
\alias{capacity<-,tiledb_array_schema-method}
\title{Sets the schema capacity (for sparse fragments)}
\usage{
capacity(x) <- value

\S4method{capacity}{tiledb_array_schema}(x) <- value
}
\arguments{
\item{x}{An \code{array_schema} object}

\item{value}{An integer or numeric value for the new tile capacity}
}
\value{
The modified \code{array_schema} object
}
\description{
Sets the \code{tiledb_array} schema tile capacity for sparse fragments.
}
