% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{close.tiledb_dense}
\alias{close.tiledb_dense}
\title{Closes a tiledb array object}
\usage{
\method{close}{tiledb_dense}(conn, ...)
}
\arguments{
\item{conn}{tiledb array object}
}
\value{
returns the closed array object
}
\description{
Closes a tiledb array object
}
