% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\docType{methods}
\name{compressor,tiledb_array_schema-method}
\alias{compressor,tiledb_array_schema-method}
\title{Returns the offsets and coordinate compressors associated with the \code{tiledb_array_schema}}
\usage{
\S4method{compressor}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{tiledb_array_schema}
}
\value{
a list of tiledb_compressor objects
}
\description{
Returns the offsets and coordinate compressors associated with the \code{tiledb_array_schema}
}
