% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{methods}
\name{compressor,tiledb_attr-method}
\alias{compressor,tiledb_attr-method}
\title{Returns the \code{tiledb_compressor} object associated with the given \code{tiledb_attr}}
\usage{
\S4method{compressor}{tiledb_attr}(object)
}
\arguments{
\item{object}{tiledb_attr}
}
\description{
Returns the \code{tiledb_compressor} object associated with the given \code{tiledb_attr}
}
