% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compressor.R
\docType{methods}
\name{compressor_level,tiledb_compressor-method}
\alias{compressor_level,tiledb_compressor-method}
\title{Returns the compressor's compression level}
\usage{
\S4method{compressor_level}{tiledb_compressor}(object)
}
\arguments{
\item{object}{tiledb_compressor}
}
\value{
integer level
}
\description{
Returns the compressor's compression level
}
\examples{
c <- tiledb_compressor("ZSTD", 3)
compressor_level(c)

}
