% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compressor.R
\docType{methods}
\name{compressor_type,tiledb_compressor-method}
\alias{compressor_type,tiledb_compressor-method}
\title{Returns the compressor's compression type}
\usage{
\S4method{compressor_type}{tiledb_compressor}(object)
}
\arguments{
\item{object}{tiledb_compressor}
}
\value{
TileDB compression type string
}
\description{
Compression library used
}
\examples{
c <- tiledb_compressor("ZSTD", 3)
compressor_type(c)

}
