% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_tiledb_array}
\alias{create_tiledb_array}
\title{All-in-one function to create a TileDB array}
\usage{
create_tiledb_array(array_name, dims, type, sparse = FALSE,
  ncells = rep.int(1, length(type)))
}
\arguments{
\item{array_name}{single character, tne name of the TileDB array file to be created}

\item{dims}{integer or numeric, desired dimensions of an attribute, to be repeated for
every attribute, e.g. c(2,3) for a 2 row, 3 column matrix}

\item{type}{character, one value per attribute}

\item{sparse}{single logical, will the created array be sparse?}

\item{ncells}{integer, one per attribute. Use 1 for fixed-length arrays and -1
for variable length arrays (including character arrays).}
}
\value{
a TileDB array object with the specified properties
}
\description{
Creates a dense or sparse array using sensible defaults.
}
\examples{
filename = tempfile()
create_tiledb_array(filename, c(2,3), "FLOAT64")
filename = tempfile()
create_tiledb_array(filename, c(2,3), c("FLOAT64", "FLOAT64"))
}
