% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchedQuery.R
\name{createBatched}
\alias{createBatched}
\title{Create a \sQuote{batched} query object}
\usage{
createBatched(x)
}
\arguments{
\item{x}{A \code{tiledb_array} object}
}
\value{
A \code{batchedquery} object, that is a list containing an external pointer
to a TileDB Query object along with other support variables used by \code{fetchBatched}
}
\description{
Batched queries return an initial result set even when it is incomplete. Where
the normal retrieval process will loop in place to complete a (potentially large)
result set, this function will return a result (which may be part of a larger
result set) allowing the user to assemble all part.
}
\details{
The \code{tiledb_array} object can be parameterised as usual.
}
