% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{data.frame.conversion<-}
\alias{data.frame.conversion<-}
\alias{data.frame.conversion<-,tiledb_array-method}
\title{Retrieve data.frame conversion preference}
\usage{
data.frame.conversion(x) <- value

\S4method{data.frame.conversion}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A character value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can be returned as a data.frame (or, if selected,
as a \code{matrix} or as an \code{array}. This methods permits to set a preference
of returning a \code{data.frame}, a \code{data.table}, or a \code{tibble}. The default
value of \dQuote{none} means that no conversion is performed.
}
