% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{data.frame.conversion}
\alias{data.frame.conversion}
\alias{data.frame.conversion,tiledb_array-method}
\title{Retrieve data.frame conversion preference}
\usage{
data.frame.conversion(object, ...)

\S4method{data.frame.conversion}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}

\item{...}{Currently unused}
}
\value{
A character value indicating the preferred conversion where the value is
one \dQuote{none} (the default), \dQuote{data.frame}, \dQuote{data.table}, or
\dQuote{tibble}.
}
\description{
A \code{tiledb_array} object can be returned as a data.frame (or, if selected,
as a \code{matrix} or as an \code{array}. This methods permits to select a preference
of returning a \code{data.frame}, a \code{data.table}, or a \code{tibble}. The default
value of \dQuote{none} means that no conversion is performed.
}
