% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{methods}
\name{datatype,tiledb_attr-method}
\alias{datatype,tiledb_attr-method}
\title{Return the \code{tiledb_attr} datatype}
\usage{
\S4method{datatype}{tiledb_attr}(object)
}
\arguments{
\item{`tiledb_attr`}{object}

\item{tiledb}{datatype string}
}
\description{
Return the \code{tiledb_attr} datatype
}
\examples{
a1 <- tiledb_attr("a1", type = "INT32")
datatype(a1)

a2 <- tiledb_attr("a1", type = "FLOAT64")
datatype(a2)

}
