% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\docType{methods}
\name{datatype,tiledb_dim-method}
\alias{datatype,tiledb_dim-method}
\title{Return the \code{tiledb_dim} tile extent}
\usage{
\S4method{datatype}{tiledb_dim}(object)
}
\arguments{
\item{`tiledb_dim`}{object}

\item{tiledb}{datatype string}
}
\description{
Return the \code{tiledb_dim} tile extent
}
\examples{
ctx <- tiledb_ctx()
d1 <- tiledb_dim(ctx, "d1", domain = c(5L, 10L), tile = 2L, type = "INT32")
datatype(d1)

}
