% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\docType{methods}
\name{datatype,tiledb_domain-method}
\alias{datatype,tiledb_domain-method}
\title{Returns tiledb_domain TileDB type string}
\usage{
\S4method{datatype}{tiledb_domain}(object)
}
\arguments{
\item{object}{tiledb_domain}
}
\value{
tiledb_domain type string
}
\description{
Returns tiledb_domain TileDB type string
}
\examples{
ctx <- tiledb_ctx()
dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(1L, 100L), type = "INT32")))
datatype(dom)

dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(0.5, 100.0), type = "FLOAT64")))
datatype(dom)

}
