% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{datetimes_as_int64<-}
\alias{datetimes_as_int64<-}
\alias{datetimes_as_int64<-,tiledb_array-method}
\title{Set datetimes_as_int64 toggle}
\usage{
datetimes_as_int64(x) <- value

\S4method{datetimes_as_int64}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object mayb contain date and datetime objects. While their internal
representation is generally shielded from the user, it can useful to access them as the
\sQuote{native} format which is an \code{integer64}. This function set the current
value of the selection variable, which has a default of \code{FALSE}.
}
