% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{datetimes_as_int64}
\alias{datetimes_as_int64}
\alias{datetimes_as_int64,tiledb_array-method}
\title{Retrieve datetimes_as_int64 toggle}
\usage{
datetimes_as_int64(object)

\S4method{datetimes_as_int64}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A logical value indicating whether \code{datetimes_as_int64} is selected
}
\description{
A \code{tiledb_array} object mayb contain date and datetime objects. While their internal
representation is generally shielded from the user, it can useful to access them as the
\sQuote{native} format which is an \code{integer64}. This function retrieves the current
value of the selection variable, which has a default of \code{FALSE}.
}
