% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{dim.tiledb_dim}
\alias{dim.tiledb_dim}
\title{Retrieves the dimension of the tiledb_dim domain}
\usage{
\method{dim}{tiledb_dim}(x)
}
\arguments{
\item{x}{\code{tiledb_dim} object}
}
\value{
a vector of the tile_dim domain type, of the dim domain dimension (extent)
}
\description{
Retrieves the dimension of the tiledb_dim domain
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
d1 <- tiledb_dim("d1", c(1L, 10L), 5L)
dim(d1)

}
