% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{dim.tiledb_domain}
\alias{dim.tiledb_domain}
\title{Retrieve the dimension (domain extent) of the domain}
\usage{
\method{dim}{tiledb_domain}(x)
}
\arguments{
\item{x}{tiledb_domain}
}
\value{
dimension vector
}
\description{
Only valid for integral (integer) domains
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 100L), type = "INT32"),
                              tiledb_dim("d2", c(1L, 100L), type = "INT32")))
dim(dom)

}
