% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\docType{methods}
\name{dimensions,tiledb_domain-method}
\alias{dimensions,tiledb_domain-method}
\title{Returns a list of the tiledb_domain dimension objects}
\usage{
\S4method{dimensions}{tiledb_domain}(object)
}
\arguments{
\item{object}{tiledb_domain}
}
\value{
a list of tiledb_dim
}
\description{
Returns a list of the tiledb_domain dimension objects
}
\examples{
ctx <- tiledb_ctx()
dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(1L, 100L), type = "INT32"),
                                   tiledb_dim(ctx, "d2", c(1L, 50L), type = "INT32")))
dimensions(dom)

lapply(dimensions(dom), name)

}
