% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\docType{methods}
\name{domain,tiledb_dim-method}
\alias{domain,tiledb_dim-method}
\title{Return the \code{tiledb_dim} domain}
\usage{
\S4method{domain}{tiledb_dim}(object)
}
\arguments{
\item{`tiledb_dim`}{object}

\item{a}{vector of (lb, ub) inclusive domain of the dimension}
}
\description{
Return the \code{tiledb_dim} domain
}
\examples{
ctx <- tiledb_ctx()
d1 <- tiledb_dim(ctx, "d1", domain = c(5L, 10L))
domain(d1)

}
