% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArrayPiping.R
\name{.parse_query_condition}
\alias{.parse_query_condition}
\title{Create a 'tiledb_query_condition' object for from an expression}
\usage{
.parse_query_condition(ta, expr, debug = TRUE, strict = TRUE)
}
\arguments{
\item{expr}{An expression that is understood by the TileDB grammar for
query conditions.}

\item{debug}{A boolean toogle to enable more verbose operations, defaults
to 'FALSE'.}

\item{strict}{A boolean toogle to enable more stringent operations, i.e. stopping
on error rather than carrying one, defaults to 'TRUE'.}
}
\value{
A \code{tiledb_query_condition} object
}
\description{
The grammar for query conditions is at present constraint to six operators
and three boolean types.
}
