% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.R
\name{.pkg_config}
\alias{.pkg_config}
\title{Compiler Arguments for Using \code{libtiledb}}
\usage{
.pkg_config(opt = c("PKG_CXX_FLAGS", "PKG_CXX_LIBS"))
}
\arguments{
\item{opt}{A single character value with the package configuration variable
to fetch; choose from
\itemize{
\item \dQuote{\code{PKG_CXX_FLAGS}}: compiler flags for \code{libtiledb}
\item \dQuote{\code{PKG_CXX_LIBS}}: linking flags for \code{libtiledb}
}}
}
\value{
A single string containing either the include directories or linking
directories for \code{libtiledb}
}
\description{
Compiler Arguments for Using \code{libtiledb}
}
\examples{
.pkg_config()
.pkg_config("PKG_CXX_LIBS")

}
\keyword{internal}
