% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{extended}
\alias{extended}
\alias{extended,tiledb_array-method}
\title{Retrieve data.frame extended returns columns toggle}
\usage{
extended(object)

\S4method{extended}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A logical value indicating whether an \code{extended} return is selected
}
\description{
A \code{tiledb_array} object can be returned as \code{data.frame}. This methods
returns the selection value for \sQuote{extended} format including row (and column,
if present) indices.
}
