% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchedQuery.R
\name{fetchBatched}
\alias{fetchBatched}
\title{Run a \sQuote{batched} query}
\usage{
fetchBatched(x, obj)
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{obj}{A \code{batchedquery} object as returned by \code{createBatched}}
}
\value{
A data.frame object with the (potentially partial) result of a
batched query
}
\description{
Batched queries return an initial result set even when it is incomplete. Where
the normal retrieval process will loop in place to complete a (potentially large)
result set, this function will return a result (which may be part of a larger
result set) allowing the user to assemble all part.
}
\details{
The \code{tiledb_array} object can be parameterised as usual.
}
