% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{filter_list<-,tiledb_attr-method}
\alias{filter_list<-,tiledb_attr-method}
\title{Sets the TileDB Filter List for the TileDB Attribute object}
\usage{
\S4method{filter_list}{tiledb_attr}(x) <- value
}
\arguments{
\item{x}{TileDB Attribute}

\item{value}{TileDB Filter List}
}
\value{
The modified TileDB Attribute object
}
\description{
Sets the TileDB Filter List for the TileDB Attribute object
}
