% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{filter_list<-,tiledb_dim-method}
\alias{filter_list<-,tiledb_dim-method}
\title{Sets the TileDB Filter List for the TileDB Dimension object}
\usage{
\S4method{filter_list}{tiledb_dim}(x) <- value
}
\arguments{
\item{x}{TileDB Dimension}

\item{value}{TileDB Filter List}
}
\value{
The modified TileDB Dimension object
}
\description{
Sets the TileDB Filter List for the TileDB Dimension object
}
