% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{filter_list,tiledb_dim-method}
\alias{filter_list,tiledb_dim-method}
\title{Returns the TileDB Filter List object associated with the given TileDB Dimension}
\usage{
\S4method{filter_list}{tiledb_dim}(object)
}
\arguments{
\item{object}{TileDB_Dimension}
}
\value{
A TileDB_filter_list object
}
\description{
Returns the TileDB Filter List object associated with the given TileDB Dimension
}
