% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix.R
\name{fromMatrix}
\alias{fromMatrix}
\alias{toMatrix}
\title{Create a TileDB array from an R matrix, or return an R matrix}
\usage{
fromMatrix(obj, uri, filter = "ZSTD", capacity = 10000L)

toMatrix(uri)
}
\arguments{
\item{obj}{A sparse matrix object.}

\item{uri}{A character variable with an Array URI.}

\item{filter}{A character variable vector, defaults to \sQuote{ZSTD}, for
one or more filters to be applied to each attribute;}

\item{capacity}{A integer value with the schema capacity, default is 10000.}
}
\value{
Null, invisibly.
}
\description{
The functions \code{fromMatrix} and \code{toMatrix} help in storing
(and retrieving) matrices using a TileDB backend. In particular they help
for matrices with explicit rownames.
}
