% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseMatrix.R
\name{fromSparseMatrix}
\alias{fromSparseMatrix}
\alias{toSparseMatrix}
\title{Create a TileDB sparse array from a given sparse matrix object}
\usage{
fromSparseMatrix(
  obj,
  uri,
  cell_order = "ROW_MAJOR",
  tile_order = "ROW_MAJOR",
  filter = "ZSTD",
  capacity = 10000L
)

toSparseMatrix(uri)
}
\arguments{
\item{obj}{A sparse matrix object.}

\item{uri}{A character variable with an Array URI.}

\item{cell_order}{A character variable with one of the TileDB cell order values,
default is \dQuote{COL_MAJOR}.}

\item{tile_order}{A character variable with one of the TileDB tile order values,
default is \dQuote{COL_MAJOR}.}

\item{filter}{A character variable vector, defaults to \sQuote{ZSTD}, for
one or more filters to be applied to each attribute;}

\item{capacity}{A integer value with the schema capacity, default is 1000.}
}
\value{
Null, invisibly.
}
\description{
Create a TileDB sparse array from a given sparse matrix object
}
