% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{getContext}
\alias{getContext}
\title{Retrieve a TileDB context object from the package cache}
\usage{
getContext()
}
\value{
A TileDB context object
}
\description{
Retrieve a TileDB context object from the package cache
}
