% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{has_attribute}
\alias{has_attribute}
\title{Check a schema for a given attribute name}
\usage{
has_attribute(schema, attr)
}
\arguments{
\item{schema}{A schema for a TileDB Array}

\item{attr}{A character variable with an attribute name}
}
\value{
A boolean value indicating if the attribute exists in the schema
}
\description{
Check a schema for a given attribute name
}
