% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{is.anonymous}
\alias{is.anonymous}
\alias{is.anonymous.tiledb_attr}
\title{Returns TRUE if the tiledb_dim is anonymous}
\usage{
is.anonymous(object)

\method{is.anonymous}{tiledb_attr}(object)
}
\arguments{
\item{object}{\code{tiledb_attr} object}
}
\value{
TRUE or FALSE
}
\description{
A TileDB attribute is anonymous if no name/label is defined
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
a1 <- tiledb_attr("a1", type = "FLOAT64")
is.anonymous(a1)

a2 <- tiledb_attr("", type = "FLOAT64")
is.anonymous(a2)

}
