% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{is.anonymous.tiledb_attr}
\alias{is.anonymous.tiledb_attr}
\title{Returns TRUE if the tiledb_dim is anonymous}
\usage{
\method{is.anonymous}{tiledb_attr}(object)
}
\arguments{
\item{`tiledb_attr`}{object}
}
\value{
TRUE or FALSE
}
\description{
A TileDB attribute is anonymous if no name/label is defined
}
\examples{
ctx <- tiledb_ctx()
a1 <- tiledb_attr(ctx, "a1", type = "FLOAT64")
is.anonymous(a1)

d2 <- tiledb_attr(ctx, "", type = "FLOAT64")
is.anonymous(a2)

}
