% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{is.anonymous.tiledb_dim}
\alias{is.anonymous.tiledb_dim}
\title{Returns TRUE if the tiledb_dim is anonymous}
\usage{
\method{is.anonymous}{tiledb_dim}(object)
}
\arguments{
\item{object}{\code{tiledb_dim} object}
}
\value{
TRUE or FALSE
}
\description{
A TileDB dimension is anonymous if no name/label is defined
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
d1 <- tiledb_dim("d1", c(1L, 10L), 10L)
is.anonymous(d1)

d2 <- tiledb_dim("", c(1L, 10L), 10L)
is.anonymous(d2)

}
