% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{is.integral,tiledb_domain-method}
\alias{is.integral,tiledb_domain-method}
\title{Returns TRUE is tiledb_domain is an integral (integer) domain}
\usage{
\S4method{is.integral}{tiledb_domain}(object)
}
\arguments{
\item{object}{tiledb_domain}
}
\value{
TRUE if the domain is an integral domain, else FALSE
}
\description{
Returns TRUE is tiledb_domain is an integral (integer) domain
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 100L), type = "INT32")))
is.integral(dom)
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(0.5, 100.0), type = "FLOAT64")))
is.integral(dom)

}
