% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{is.sparse,tiledb_array_schema-method}
\alias{is.sparse,tiledb_array_schema-method}
\title{Returns TRUE if the tiledb_array_schema is sparse, else FALSE}
\usage{
\S4method{is.sparse}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{A TileDB Schema object}
}
\value{
\code{TRUE} if \code{tiledb_array_schema} object is sparse
}
\description{
Returns TRUE if the tiledb_array_schema is sparse, else FALSE
}
