% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{is.sparse,tiledb_sparse-method}
\alias{is.sparse,tiledb_sparse-method}
\title{Check if object is sparse}
\usage{
\S4method{is.sparse}{tiledb_sparse}(object)
}
\arguments{
\item{object}{TileDB object}
}
\value{
A logical value indicating whether the object is sparse
}
\description{
Check if object is sparse
}
