% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\docType{methods}
\name{max_chunk_size,tiledb_filter_list-method}
\alias{max_chunk_size,tiledb_filter_list-method}
\title{Returns the filter_list's max_chunk_size}
\usage{
\S4method{max_chunk_size}{tiledb_filter_list}(object)
}
\arguments{
\item{object}{tiledb_filter_list}
}
\value{
integer max_chunk_size
}
\description{
Returns the filter_list's max_chunk_size
}
\examples{
flt = tiledb_filter(ctx, "ZSTD")
tiledb_filter_set_option(flt, "COMPRESSION_LEVEL", 5)
flt.set
filter_list <- tiledb_filter_list(ctx, c(flt))
max_chunk_size(filter_list)

}
