% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{name,tiledb_dim-method}
\alias{name,tiledb_dim-method}
\title{Return the \code{tiledb_dim} name}
\usage{
\S4method{name}{tiledb_dim}(object)
}
\arguments{
\item{object}{\code{tiledb_dim} object}
}
\value{
string name, empty string if the dimension is anonymous
}
\description{
Return the \code{tiledb_dim} name
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
d1 <- tiledb_dim("d1", c(1L, 10L))
name(d1)

d2 <- tiledb_dim("", c(1L, 10L))
name(d2)

}
