% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{ncells,tiledb_attr-method}
\alias{ncells,tiledb_attr-method}
\title{Return the number of scalar values per attribute cell}
\usage{
\S4method{ncells}{tiledb_attr}(object)
}
\arguments{
\item{`tiledb_attr`}{object}
}
\value{
integer number of cells
}
\description{
Return the number of scalar values per attribute cell
}
\examples{
a1 <- tiledb_attr("a1", type = "FLOAT64", ncells = 1)
ncells(a1)

}
