% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\name{nfilters,tiledb_filter_list-method}
\alias{nfilters,tiledb_filter_list-method}
\title{Returns the filter_list's number of filters}
\usage{
\S4method{nfilters}{tiledb_filter_list}(object)
}
\arguments{
\item{object}{tiledb_filter_list}
}
\value{
integer number of filters
}
\description{
Returns the filter_list's number of filters
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
flt <- tiledb_filter("ZSTD")
tiledb_filter_set_option(flt, "COMPRESSION_LEVEL", 5)
filter_list <- tiledb_filter_list(c(flt))
nfilters(filter_list)

}
