% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{print.tiledb_metadata}
\alias{print.tiledb_metadata}
\title{Print a TileDB Array Metadata object}
\usage{
\method{print}{tiledb_metadata}(x, width = NULL, ...)
}
\arguments{
\item{x}{A TileDB array object}

\item{width}{Optional display width, defaults to NULL}

\item{...}{Optional method arguments, currently unused}
}
\value{
The array object, invisibly
}
\description{
Print a TileDB Array Metadata object
}
