% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{query_condition<-}
\alias{query_condition<-}
\alias{query_condition<-,tiledb_array-method}
\title{Set query_condition object for the array}
\usage{
query_condition(x) <- value

\S4method{query_condition}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A \code{tiledb_query_conditon_object}}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can have an associated query condition object to set
conditions on the read queries. This methods sets the \sQuote{query_condition} object.
}
