% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{query_condition}
\alias{query_condition}
\alias{query_condition,tiledb_array-method}
\title{Retrieve query_condition value for the array}
\usage{
query_condition(object)

\S4method{query_condition}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A \code{tiledb_query_condition} object
}
\description{
A \code{tiledb_array} object can have a corresponding query condition object.
This methods returns it.
}
