% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{query_layout}
\alias{query_layout}
\alias{query_layout,tiledb_array-method}
\title{Retrieve query_layout values for the array}
\usage{
query_layout(object)

\S4method{query_layout}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A character value describing the query layout
}
\description{
A \code{tiledb_array} object can have a corresponding query with a given layout
given layout. This methods returns the selection value for \sQuote{query_layout}
as a character value.
}
