% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{query_statistics<-}
\alias{query_statistics<-}
\alias{query_statistics<-,tiledb_array-method}
\title{Set query_statistics toggle}
\usage{
query_statistics(x) <- value

\S4method{query_statistics}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can, if requested, return query statistics as a JSON
string in an attribute \sQuote{query_statistics} attached to the return object. The
default value of the logical switch is \sQuote{FALSE}. This method sets the value.
}
