% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{query_statistics}
\alias{query_statistics}
\alias{query_statistics,tiledb_array-method}
\title{Retrieve query_statistics toggle}
\usage{
query_statistics(object, ...)

\S4method{query_statistics}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}

\item{...}{Currently unused}
}
\value{
A logical value indicating whether query statistics are returned.
}
\description{
A \code{tiledb_array} object can, if requested, return query statistics as a JSON
string in an attribute \sQuote{query_statistics} attached to the return object. The
default value of the logical switch is \sQuote{FALSE}. This method returns the current
value.
}
