% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{r_to_tiledb_type}
\alias{r_to_tiledb_type}
\title{Look up TileDB type corresponding to the type of an R object}
\usage{
r_to_tiledb_type(x)
}
\arguments{
\item{x}{an R array or list}
}
\value{
single character, e.g. INT32
}
\description{
Look up TileDB type corresponding to the type of an R object
}
