% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{raw_dump,tiledb_domain-method}
\alias{raw_dump,tiledb_domain-method}
\title{Raw display of a domain object}
\usage{
\S4method{raw_dump}{tiledb_domain}(object)
}
\arguments{
\item{object}{A domain object}
}
\description{
This method used the display method provided by the underlying
library.
}
