% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{raw_dump,tiledb_profile-method}
\alias{raw_dump,tiledb_profile-method}
\title{Raw display of a profile object}
\usage{
\S4method{raw_dump}{tiledb_profile}(object)
}
\arguments{
\item{object}{A profile object}
}
\description{
This method uses the display method provided by the underlying library.
}
