% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableLength.R
\name{read_variable_length}
\alias{read_variable_length}
\title{Read an variable-length array}
\usage{
read_variable_length(uri, key, subarray, debug = FALSE)
}
\arguments{
\item{uri}{A character variable with a URI for a (dense) array}

\item{key}{A character variable with the array key}

\item{subarray}{An integer vector of length four with the row start and
and end, and the column start and and end (using one as start).}

\item{debug}{An optional logical switch for increased verbosity, defaults
to FALSE.}
}
\value{
A data.table object is returned if the data.table package is available,
otherwise a list object is returned.
}
\description{
Using a given key and sub-array vector, a variable length array is read.
Currently defaults to dense arrays.
}
