% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{reopen}
\alias{reopen}
\title{Reopens a TileDB array an opened tiledb array}
\usage{
reopen(object, ...)
}
\arguments{
\item{object}{tileb array object}
}
\value{
the reopened array object
}
\description{
Reopening an array is useful when the array got updated after it got opened
and the tiledb array object got created. To sync-up with the updates,
the user must either close the array and open again,
or just use tiledb_reopen(array) which can be faster because
only metdata regarding updates has to be loaded.
}
