% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return.array<-}
\alias{return.array<-}
\alias{return.array<-,tiledb_array-method}
\title{Set array return toggle}
\usage{
return.array(x) <- value

\S4method{return.array}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame} or a \code{matrix}. This methods sets the
selection value for a \code{array}.
}
