% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return.array}
\alias{return.array}
\alias{return.array,tiledb_array-method}
\title{Retrieve array return toggle}
\usage{
return.array(object, ...)

\S4method{return.array}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}

\item{...}{Currently unused}
}
\value{
A logical value indicating whether \code{array} return is selected
}
\description{
A \code{tiledb_array} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame} or as a \code{matrix}. This methods returns
the selection value for the \code{array} selection.
}
