% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{return.data.frame<-}
\alias{return.data.frame<-}
\alias{return.data.frame<-,tiledb_dense-method}
\title{Set data.frame return toggle}
\usage{
return.data.frame(x) <- value

\S4method{return.data.frame}{tiledb_dense}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_dense} array object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_dense} array object
}
\description{
A \code{tiledb_dense} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame}. This methods sets the selection value.
}
