% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{return.data.frame<-,tiledb_sparse-method}
\alias{return.data.frame<-,tiledb_sparse-method}
\title{Set data.frame return toggle}
\usage{
\S4method{return.data.frame}{tiledb_sparse}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_sparse} array object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_sparse} array object
}
\description{
A \code{tiledb_sparse} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame}. This methods sets the selection value.
}
