% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return.data.frame,tiledb_array-method}
\alias{return.data.frame,tiledb_array-method}
\title{Retrieve data.frame return toggle}
\usage{
\S4method{return.data.frame}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A logical value indicating whether \code{data.frame} return is selected
}
\description{
A \code{tiledb_array} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame}. This methods returns the selection value.
}
