% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{return.data.frame}
\alias{return.data.frame}
\alias{return.data.frame,tiledb_dense-method}
\title{Retrieve data.frame return toggle}
\usage{
return.data.frame(object, ...)

\S4method{return.data.frame}{tiledb_dense}(object)
}
\arguments{
\item{object}{A \code{tiledb_dense} array object}

\item{...}{Currently unused}
}
\value{
A logical value indicating whether \code{data.frame} return is selected
}
\description{
A \code{tiledb_dense} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame}. This methods returns the selection value.
}
