% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{return.data.frame,tiledb_sparse-method}
\alias{return.data.frame,tiledb_sparse-method}
\title{Retrieve data.frame return toggle}
\usage{
\S4method{return.data.frame}{tiledb_sparse}(object)
}
\arguments{
\item{object}{A \code{tiledb_sparse} array object}
}
\value{
A logical value indicating whether \code{data.frame} return is selected
}
\description{
A \code{tiledb_sparse} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame}. This methods returns the selection value.
}
