% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return.matrix}
\alias{return.matrix}
\alias{return.matrix,tiledb_array-method}
\title{Retrieve matrix.frame return toggle}
\usage{
return.matrix(object, ...)

\S4method{return.matrix}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}

\item{...}{Currently unused}
}
\value{
A logical value indicating whether \code{matrix} return is selected
}
\description{
A \code{tiledb_array} object can be returned as an array (or list of arrays),
or, if select, as a \code{data.frame} or as a \code{matrix}. This methods returns
the selection value for the \code{matrix} selection.
}
