% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return_as<-}
\alias{return_as<-}
\alias{return_as<-,tiledb_array-method}
\title{Retrieve return_as conversion preference}
\usage{
return_as(x) <- value

\S4method{return_as}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A character value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can be returned as a \sQuote{list} (default), \sQuote{array},
\sQuote{matrix}, \sQuote{data.frame}, \sQuote{data.table} or \sQuote{tibble}. This method
This methods permits to set a preference of returning a \code{list}, \code{array},
\code{matrix}, \code{data.frame}, a \code{data.table}, or a \code{tibble}. The default
value of \dQuote{asis} means that no conversion is performed and a \code{list} is returned.
}
