% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{return_as}
\alias{return_as}
\alias{return_as,tiledb_array-method}
\title{Retrieve return_as conversion preference}
\usage{
return_as(object, ...)

\S4method{return_as}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}

\item{...}{Currently unused}
}
\value{
A character value indicating the preferred conversion where the value is
one of \sQuote{asis} (the default), \sQuote{array}, \sQuote{matrix},\sQuote{data.frame},
\sQuote{data.table}, or \sQuote{tibble}.
}
\description{
A \code{tiledb_array} object can be returned as a \sQuote{list} (default), \sQuote{array},
\sQuote{matrix}, \sQuote{data.frame}, \sQuote{data.table} or \sQuote{tibble}. This method
permits to select a preference for the returned object. The default value of \sQuote{asis}
means that no conversion is performed.
}
