% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{save_dataframe_conversion_preference}
\alias{save_dataframe_conversion_preference}
\alias{load_dataframe_conversion_preference}
\alias{get_dataframe_conversion_preference}
\alias{set_dataframe_conversion_preference}
\title{Store data.frame conversion preference}
\usage{
save_dataframe_conversion_preference(
  value = c("none", "data.frame", "data.table", "tibble")
)

load_dataframe_conversion_preference()

get_dataframe_conversion_preference()

set_dataframe_conversion_preference(
  value = c("none", "data.frame", "data.table", "tibble")
)
}
\arguments{
\item{value}{A character variable with one of the four permitted values}
}
\value{
For the setter, \code{TRUE} is returned invisibly but the function is invoked for the
side effect of storing the value. For either getter, the character value.
}
\description{
Save (or load) data.frame conversion preference in an optional config file
}
\details{
The \code{tiledb_array} object can set a preference for \code{data.frame}
conversion for each retrieved object. This preference can also be enconded
in configuration file as R (version 4.0.0 or later) allows a user- and
package specific configuration file.  These helper functions sets and retrieve
the value, respectively, or retrieve the cached value from the package environment
where is it set at package load.

Note that the value must be one of \dQuote{none} (the default), \dQuote{data.frame},
\dQuote{data.table} or \dQuote{tibble}.
}
\note{
This function requires R version 4.0.0 or later to utilise the per-user
config directory accessor function. For older R versions, please set the attribute
directly when creating the \code{tiledb_array} object, or via the
\code{data.frame_conversion()} method.
}
