% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{schema,character-method}
\alias{schema,character-method}
\title{Return a schema from a URI character value}
\usage{
\S4method{schema}{character}(object, ...)
}
\arguments{
\item{object}{A character variable with a URI}

\item{...}{Extra parameters such as \sQuote{enckey}, the encryption key}
}
\value{
The scheme for the object
}
\description{
Return a schema from a URI character value
}
