% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{schema,tiledb_array-method}
\alias{schema,tiledb_array-method}
\title{Return a schema from a tiledb_array object}
\usage{
\S4method{schema}{tiledb_array}(object, ...)
}
\arguments{
\item{object}{tiledb array object}

\item{...}{Extra parameter for function signature, currently unused}
}
\value{
The scheme for the object
}
\description{
Return a schema from a tiledb_array object
}
