% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{schema,tiledb_dense-method}
\alias{schema,tiledb_dense-method}
\title{Returns the \code{tiledb_dense} array \code{tiledb_schema} object}
\usage{
\S4method{schema}{tiledb_dense}(object, ...)
}
\arguments{
\item{object}{tiledb_dense array object}

\item{...}{Extra parameter for method signature, currently unused.}
}
\value{
tiledb_schema
}
\description{
Returns the \code{tiledb_dense} array \code{tiledb_schema} object
}
