% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{schema,tiledb_sparse-method}
\alias{schema,tiledb_sparse-method}
\title{Return a schema from a sparse array}
\usage{
\S4method{schema}{tiledb_sparse}(object, ...)
}
\arguments{
\item{object}{sparse array object}

\item{...}{Extra parameter for function signature, currently unused}
}
\value{
The scheme for the object
}
\description{
Return a schema from a sparse array
}
