% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbArray.R
\name{schema,tiledbArray-method}
\alias{schema,tiledbArray-method}
\title{Return a schema from a tiledbArray object}
\usage{
\S4method{schema}{tiledbArray}(object, ...)
}
\arguments{
\item{object}{sparse array object}

\item{...}{Extra parameter for function signature, currently unused}
}
\value{
The scheme for the object
}
\description{
Return a schema from a tiledbArray object
}
