% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{selected_points<-}
\alias{selected_points<-}
\alias{selected_points<-,tiledb_array-method}
\title{Set selected_points return values for the array}
\usage{
selected_points(x) <- value

\S4method{selected_points}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A list of vectors where each list element \sQuote{i}
corresponds to the dimension attribute \sQuote{i}.}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can have a range selection for each dimension
attribute. This methods sets the selection value for \sQuote{selected_points}
which is a list (with one element per dimension) of two-column matrices where
each row describes one pair of minimum and maximum values. Alternatively, the list
can be named with the names providing the match to the corresponding dimension.
}
