% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{selected_ranges<-}
\alias{selected_ranges<-}
\alias{selected_ranges<-,tiledb_array-method}
\title{Set selected_ranges return values for the array}
\usage{
selected_ranges(x) <- value

\S4method{selected_ranges}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A list of two-column matrices where each list element \sQuote{i}
corresponds to the dimension attribute \sQuote{i}. The matrices can contain rows
where each row contains the minimum and maximum value of a range.}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object can have a range selection for each dimension
attribute. This methods sets the selection value for \sQuote{selected_ranges}
which is a list (with one element per dimension) of two-column matrices where
each row describes one pair of minimum and maximum values.
}
