% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{selected_ranges}
\alias{selected_ranges}
\alias{selected_ranges,tiledb_array-method}
\title{Retrieve selected_ranges values for the array}
\usage{
selected_ranges(object)

\S4method{selected_ranges}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A list which can contain a matrix for each dimension
}
\description{
A \code{tiledb_array} object can have a range selection for each dimension
attribute. This methods returns the selection value for \sQuote{selected_ranges}
and returns a list (with one element per dimension) of two-column matrices where
each row describes one pair of minimum and maximum values. Alternatively, the list
can be named with the names providing the match to the corresponding dimension.
}
