% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{setContext}
\alias{setContext}
\title{Store a TileDB context object in the package cache}
\usage{
setContext(ctx)
}
\arguments{
\item{ctx}{A TileDB context object}
}
\value{
A TileDB context object
}
\description{
Store a TileDB context object in the package cache
}
