% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{show,tiledb_array_schema-method}
\alias{show,tiledb_array_schema-method}
\title{Prints an array schema object}
\usage{
\S4method{show}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{An array_schema object}
}
\description{
Prints an array schema object
}
