% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{show,tiledb_config-method}
\alias{show,tiledb_config-method}
\title{Prints the config object to STDOUT}
\usage{
\S4method{show}{tiledb_config}(object)
}
\arguments{
\item{object}{\code{tiledb_config} object}
}
\description{
Prints the config object to STDOUT
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
cfg <- tiledb_config()
show(cfg)
}
