% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{show,tiledb_dense-method}
\alias{show,tiledb_dense-method}
\title{Prints a tiledb_dense array object}
\usage{
\S4method{show}{tiledb_dense}(object)
}
\arguments{
\item{object}{A tiledb_dense array object}
}
\description{
Prints a tiledb_dense array object
}
