% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{show,tiledb_dim-method}
\alias{show,tiledb_dim-method}
\title{Prints a dimension object}
\usage{
\S4method{show}{tiledb_dim}(object)
}
\arguments{
\item{object}{An array_schema object}
}
\description{
Prints a dimension object
}
