% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logging.R
\name{spdl_setup}
\alias{spdl_setup}
\alias{spdl_drop}
\alias{spdl_set_pattern}
\alias{spdl_set_level}
\alias{spdl_trace}
\alias{spdl_debug}
\alias{spdl_info}
\alias{spdl_warn}
\alias{spdl_error}
\alias{spdl_critical}
\title{Convenience Wrappers for 'RcppSpdlog' Logging From 'spdlog'}
\usage{
spdl_setup(name = "default", level = "warn")

spdl_drop(s)

spdl_set_pattern(s)

spdl_set_level(s)

spdl_trace(s)

spdl_debug(s)

spdl_info(s)

spdl_warn(s)

spdl_error(s)

spdl_critical(s)
}
\arguments{
\item{name}{Character value for the name of the logger instance}

\item{level}{Character value for the logging level}

\item{s}{Character value for pattern, level, or logging message}
}
\value{
Nothing is returned from these functions as they are invoked for their side-effects.
}
\description{
Several short wrappers for functions from 'RcppSpdlog' package are provided
as a convenience.  Given the potential for clashing names of common and popular
functions names it is \emph{not} recommended to import the whole package but rather
do \code{importFrom(RcppSpdlog, set_pattern)} (or maybe \code{importFrom(RcppSpdlog,
set_pattern)}). The functions defined here are local to this package and allow for
shorter simpler access via for example \code{spld_info()} to log at the \sQuote{info} level.
}
\examples{
\dontrun{
spdl_setup("exampleDemo", "warn")
spdl_info("Not seen as level 'info' below 'warn'")
spdl_warn("This warning message is seen")
spdl_set_level("info")
spdl_info("Now this informational message is seen too")
spdl_info(sprintf("Calls to sprintf() can be used \%s as we see \%d", "as well", 42L))
}
}
