% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{strings_as_factors<-}
\alias{strings_as_factors<-}
\alias{strings_as_factors<-,tiledb_array-method}
\title{Set strings_as_factors return toggle}
\usage{
strings_as_factors(x) <- value

\S4method{strings_as_factors}{tiledb_array}(x) <- value
}
\arguments{
\item{x}{A \code{tiledb_array} object}

\item{value}{A logical value with the selection}
}
\value{
The modified \code{tiledb_array} array object
}
\description{
A \code{tiledb_array} object containing character column can have those converted to
factors variables. This methods sets the selection value for \sQuote{strings_as_factors}.
}
