% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{strings_as_factors}
\alias{strings_as_factors}
\alias{strings_as_factors,tiledb_array-method}
\title{Retrieve strings_as_factors conversion toggle}
\usage{
strings_as_factors(object)

\S4method{strings_as_factors}{tiledb_array}(object)
}
\arguments{
\item{object}{A \code{tiledb_array} object}
}
\value{
A logical value indicating whether an \code{strings_as_factors} return is selected
}
\description{
A \code{tiledb_array} object containing character column can have those converted to
factors variables. This methods returns the selection value for \sQuote{strings_as_factors}.
}
