% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{[,tiledb_array,ANY-method}
\alias{[,tiledb_array,ANY-method}
\alias{[,tiledb_array}
\alias{[,tiledb_array-method}
\alias{[,tiledb_array,ANY,tiledb_array-method}
\alias{[,tiledb_array,ANY,ANY,tiledb_array-method}
\title{Returns a TileDB array, allowing for specific subset ranges.}
\usage{
\S4method{[}{tiledb_array,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{tiledb_array object}

\item{i}{optional row index expression which can be a list in which case
minimum and maximum of each list element determine a range; multiple list
elements can be used to supply multiple ranges.}

\item{j}{optional column index expression which can be a list in which case
minimum and maximum of each list element determine a range; multiple list
elements can be used to supply multiple ranges.}

\item{...}{Extra parameters for method signature, currently unused.}

\item{drop}{Optional logical switch to drop dimensions, default FALSE,
currently unused.}
}
\value{
The resulting elements in the selected format
}
\description{
Heterogeneous domains are supported, including timestamps and characters.
}
\details{
This function may still still change; the current implementation should be
considered as an initial draft.
}
