% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{[,tiledb_config,ANY-method}
\alias{[,tiledb_config,ANY-method}
\alias{[,tiledb_config}
\alias{[,tiledb_config-method}
\alias{[,tiledb_config,ANY,tiledb_config-method}
\alias{[,tiledb_config,ANY,ANY,tiledb_config-method}
\title{Gets a config parameter value}
\usage{
\S4method{[}{tiledb_config,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{\code{tiledb_config} object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{drop}{Optional logical switch to drop dimensions, default FALSE, currently unused.}
}
\value{
a config string value if parameter exists, else NA
}
\description{
Gets a config parameter value
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
cfg <- tiledb_config()
cfg["sm.tile_cache_size"]
cfg["does_not_exist"]
}
