% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{[,tiledb_config-method}
\alias{[,tiledb_config-method}
\title{Gets a config parameter value}
\usage{
\S4method{[}{tiledb_config}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\code{tiledb_config}object}

\item{i}{parameter key string}
}
\value{
a config string value if parameter exists, else NA
}
\description{
Gets a config parameter value
}
\examples{
cfg <- tiledb_config()
cfg["sm.tile_cache_size"]
cfg["does_not_exist"]

}
