% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{[,tiledb_dense,ANY-method}
\alias{[,tiledb_dense,ANY-method}
\alias{[,tiledb_dense}
\alias{[,tiledb_dense-method}
\alias{[,tiledb_dense,ANY,tiledb_dense-method}
\alias{[,tiledb_dense,ANY,ANY,tiledb_dense-method}
\title{Gets a dense array value}
\usage{
\S4method{[}{tiledb_dense,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{dense array object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{drop}{Optional logical switch to drop dimensions, default FALSE, currently unused.}
}
\value{
An element from a dense array
}
\description{
Gets a dense array value
}
