% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\name{[,tiledb_filter_list,ANY-method}
\alias{[,tiledb_filter_list,ANY-method}
\alias{[,tiledb_filter_list}
\alias{[,tiledb_filter_list-method}
\alias{[,tiledb_filter_list,ANY,tiledb_filter_list-method}
\alias{[,tiledb_filter_list,ANY,ANY,tiledb_filter_list-method}
\title{Returns the filter at given index}
\usage{
\S4method{[}{tiledb_filter_list,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{\code{tiledb_config} object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{drop}{Optional logical switch to drop dimensions, default false.}
}
\value{
object tiledb_filter
}
\description{
Returns the filter at given index
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
flt <- tiledb_filter("ZSTD")
tiledb_filter_set_option(flt, "COMPRESSION_LEVEL", 5)
filter_list <- tiledb_filter_list(c(flt))
filter_list[0]

}
