% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{[,tiledb_sparse,ANY-method}
\alias{[,tiledb_sparse,ANY-method}
\alias{[,tiledb_sparse}
\alias{[,tiledb_sparse-method}
\alias{[,tiledb_sparse,ANY,tiledb_sparse-method}
\alias{[,tiledb_sparse,ANY,ANY,tiledb_sparse-method}
\title{Gets a sparse array value}
\usage{
\S4method{[}{tiledb_sparse,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{sparse array object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{drop}{Optional logical switch to drop dimensions, default FALSE, currently unused.}
}
\value{
An element from the sparse array
}
\description{
Gets a sparse array value
}
