% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbArray.R
\name{[,tiledbArray-method}
\alias{[,tiledbArray-method}
\title{Gets a tiledbArray value}
\usage{
\S4method{[}{tiledbArray}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{tiledbArray object}

\item{i}{row index expression}

\item{j}{column index expression}

\item{...}{Extra parameter for method signature}

\item{drop}{Optional logical switch to drop dimensions, default FALSE, currently unused.}
}
\value{
An element from the sparse array
}
\description{
This function is incomplete.
}
