% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_array.R
\name{[<-,tiledb_array,ANY,ANY,ANY-method}
\alias{[<-,tiledb_array,ANY,ANY,ANY-method}
\title{Sets a tiledb array value or value range}
\usage{
\S4method{[}{tiledb_array,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{sparse array object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{value}{The value being assigned}
}
\value{
The modified object
}
\description{
Sets a tiledb array value or value range
}
