% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{[<-,tiledb_config,ANY,ANY,ANY-method}
\alias{[<-,tiledb_config,ANY,ANY,ANY-method}
\alias{[<-,tiledb_config}
\alias{[<-,tiledb_config-method}
\alias{[<-,tiledb_config,ANY,tiledb_config-method}
\alias{[<-,tiledb_config,ANY,ANY,tiledb_config-method}
\title{Sets a config parameter value}
\usage{
\S4method{[}{tiledb_config,ANY,ANY,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{\code{tiledb_config} object}

\item{i}{parameter key string}

\item{j}{parameter key string}

\item{value}{value to set, will be converted into a stringa}
}
\value{
updated \code{tiledb_config} object
}
\description{
Sets a config parameter value
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
cfg <- tiledb_config()
cfg["sm.tile_cache_size"]

# set tile cache size to custom value
cfg["sm.tile_cache_size"] <- 100
cfg["sm.tile_cache_size"]

}
