% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{[<-,tiledb_dense,ANY,ANY,ANY-method}
\alias{[<-,tiledb_dense,ANY,ANY,ANY-method}
\alias{[<-,tiledb_dense}
\alias{[<-,tiledb_dense-method}
\alias{[<-,tiledb_dense,ANY,tiledb_dense-method}
\alias{[<-,tiledb_dense,ANY,ANY,tiledb_dense-method}
\title{Sets a dense array value}
\usage{
\S4method{[}{tiledb_dense,ANY,ANY,ANY}(x, i, j, ...) <- value
}
\arguments{
\item{x}{dense array object}

\item{i}{parameter key string}

\item{j}{parameter key string, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}

\item{value}{The value being assigned}
}
\value{
The modified object
}
\description{
Sets a dense array value
}
