% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{tdb_collect,tiledb_array-method}
\alias{tdb_collect,tiledb_array-method}
\title{Collect the query results to finalize piped expression}
\usage{
\S4method{tdb_collect}{tiledb_array}(x, ...)
}
\arguments{
\item{x}{A tiledb_array object as first argument, permitting piping}

\item{...}{Ignored}
}
\value{
The object returning from a tiledb_array query (the type of which can be
set via the return preference mechanism, see the help for \code{"["} accessor)
}
\description{
Collect the query results to finalize piped expression
}
