% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{tdb_select,tiledb_array-method}
\alias{tdb_select,tiledb_array-method}
\title{Select attributes from array for query}
\usage{
\S4method{tdb_select}{tiledb_array}(x, ...)
}
\arguments{
\item{x}{A tiledb_array object as first argument, permitting piping}

\item{...}{One or more attributes of the query}
}
\value{
The tiledb_array object, permitting piping
}
\description{
Select attributes from array for query
}
