% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\docType{methods}
\name{tile,tiledb_dim-method}
\alias{tile,tiledb_dim-method}
\title{Return the \code{tiledb_dim} tile extent}
\usage{
\S4method{tile}{tiledb_dim}(object)
}
\arguments{
\item{`tiledb_dim`}{object}

\item{a}{scalar tile extent}
}
\description{
Return the \code{tiledb_dim} tile extent
}
\examples{
ctx <- tiledb_ctx()
d1 <- tiledb_dim(ctx, "d1", domain = c(5L, 10L), tile = 2L)
tile(d1)

}
