% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tile_order,tiledb_array_schema-method}
\alias{tile_order,tiledb_array_schema-method}
\title{Returns the tile layout string associated with the \code{tiledb_array_schema}}
\usage{
\S4method{tile_order}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{tiledb object}
}
\description{
Returns the tile layout string associated with the \code{tiledb_array_schema}
}
