% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiledb_array}
\alias{tiledb_array}
\title{Create a TileDB array}
\usage{
tiledb_array(x, array_name)
}
\arguments{
\item{x}{an array to fill a new TileDB array}

\item{array_name}{single character, name of new TileDB array file}
}
\value{
TileDB array object
}
\description{
Creates and fills a single-attribute, dense TileDB array using sensible defaults.
}
\examples{
x = matrix( c(1.3, 1, 2, 4), ncol = 2)
tiledb_array( x, tempfile() )
}
