% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_apply_aggregate}
\alias{tiledb_array_apply_aggregate}
\title{Run an aggregate query on the given array and attribute}
\usage{
tiledb_array_apply_aggregate(
  array,
  attrname,
  operation = c("Count", "NullCount", "Min", "Max", "Mean", "Sum"),
  nullable = TRUE
)
}
\arguments{
\item{array}{A TileDB Array object}

\item{attrname}{The name of an attribute}

\item{operation}{The name of aggregation operation}

\item{nullable}{A boolean toggle whether the attribute is nullable}
}
\value{
The value of the aggregation
}
\description{
Run an aggregate query on the given array and attribute
}
