% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_close}
\alias{tiledb_array_close}
\title{Close a TileDB Array}
\usage{
tiledb_array_close(arr)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}
}
\value{
The TileDB Array object but closed
}
\description{
Close a TileDB Array
}
