% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_delete_fragments}
\alias{tiledb_array_delete_fragments}
\title{Delete fragments written between the start and end times given}
\usage{
tiledb_array_delete_fragments(arr, ts_start, ts_end)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}

\item{ts_start}{A Datetime object that will be converted to millisecond granularity}

\item{ts_end}{A Datetime object that will be converted to millisecond granularity}
}
\value{
A boolean indicating success
}
\description{
Delete fragments written between the start and end times given
}
