% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_delete_fragments_list}
\alias{tiledb_array_delete_fragments_list}
\title{Delete fragments written given by their URIs}
\usage{
tiledb_array_delete_fragments_list(arr, fragments, ctx = tiledb_get_context())
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}

\item{fragments}{A character vector with fragment URIs}

\item{ctx}{A tiledb_ctx object (optional)}
}
\value{
A boolean indicating success
}
\description{
Delete fragments written given by their URIs
}
