% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{tiledb_array_get_non_empty_domain_from_name}
\alias{tiledb_array_get_non_empty_domain_from_name}
\title{Get the non-empty domain from a TileDB Array by name}
\usage{
tiledb_array_get_non_empty_domain_from_name(arr, name)
}
\arguments{
\item{arr}{A TileDB Array}

\item{name}{An character variable with a dimension name}
}
\value{
A two-element object is returned describing the domain of selected
dimension; it will either be a numeric vector in case of a fixed-size
fixed-sized dimensions, or a characer vector for a variable-sized one.
}
\description{
This functions works for both fixed- and variable-sized dimensions and switches
internally.
}
