% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_has_enumeration}
\alias{tiledb_array_has_enumeration}
\title{Check for Enumeration (aka Factor aka Dictionary)}
\usage{
tiledb_array_has_enumeration(arr)
}
\arguments{
\item{arr}{A TileDB Array object}
}
\value{
A logical vector indicating which attribute has enumeration
}
\description{
Check for Enumeration (aka Factor aka Dictionary)
}
