% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_is_heterogeneous}
\alias{tiledb_array_is_heterogeneous}
\title{Check for Heterogeneous Domain}
\usage{
tiledb_array_is_heterogeneous(arr)
}
\arguments{
\item{arr}{A TileDB Array object}
}
\value{
A boolean indicating if the array has heterogeneous domains
}
\description{
Check for Heterogeneous Domain
}
