% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_is_homogeneous}
\alias{tiledb_array_is_homogeneous}
\title{Check for Homogeneous Domain}
\usage{
tiledb_array_is_homogeneous(arr)
}
\arguments{
\item{arr}{A TileDB Array object}
}
\value{
A boolean indicating if the array has homogeneous domains
}
\description{
Check for Homogeneous Domain
}
