% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_is_open}
\alias{tiledb_array_is_open}
\title{Test if TileDB Array is open}
\usage{
tiledb_array_is_open(arr)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}
}
\value{
A boolean indicating whether the TileDB Array object is open
}
\description{
Test if TileDB Array is open
}
