% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_is_open_for_reading}
\alias{tiledb_array_is_open_for_reading}
\title{Test if TileDB Array is open for reading}
\usage{
tiledb_array_is_open_for_reading(arr)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}
}
\value{
A boolean indicating whether the TileDB Array object is open, and
the mode is "READ".
}
\description{
Test if TileDB Array is open for reading
}
