% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_is_open_for_writing}
\alias{tiledb_array_is_open_for_writing}
\title{Test if TileDB Array is open for writing}
\usage{
tiledb_array_is_open_for_writing(arr)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}
}
\value{
A boolean indicating whether the TileDB Array object is open, and
the mode is "WRITE".
}
\description{
Test if TileDB Array is open for writing
}
