% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_open_at}
\alias{tiledb_array_open_at}
\title{Open a TileDB Array at Timestamp}
\usage{
tiledb_array_open_at(arr, type = c("READ", "WRITE"), timestamp)
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_array()}}

\item{type}{A character value that must be either \sQuote{READ} or \sQuote{WRITE}}

\item{timestamp}{A Datetime object that will be converted to millisecond granularity}
}
\value{
The TileDB Array object but opened for reading or writing
}
\description{
Open a TileDB Array at Timestamp
}
