% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\docType{class}
\name{tiledb_array_schema-class}
\alias{tiledb_array_schema-class}
\title{An S4 class for the TileDB array schema}
\description{
An S4 class for the TileDB array schema
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}

\item{\code{arrptr}}{An optional external pointer to the underlying array, or NULL if missing}
}}

