% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema}
\alias{tiledb_array_schema}
\title{Constructs a \code{tiledb_array_schema} object}
\usage{
tiledb_array_schema(ctx, domain, attrs, cell_order = "COL_MAJOR",
  tile_order = "COL_MAJOR", sparse = FALSE, coords_compressor = NULL,
  offsets_compressor = NULL)
}
\arguments{
\item{ctx}{tiledb_ctx object}

\item{domain}{tiledb_domain object}

\item{attrs}{a list of one or more tiledb_attr objects}

\item{cell_order}{(default "COL_MAJOR")}

\item{tile_order}{(default "COL_MAJOR")}

\item{sparse}{(default FALSE)}

\item{coords_compressor}{(optional)}

\item{offsets_compressor}{(optional)}
}
\description{
Constructs a \code{tiledb_array_schema} object
}
\examples{
ctx <- tiledb_ctx()
schema <- tiledb_array_schema(ctx,
              dom = tiledb_domain(ctx, 
                        dims = c(tiledb_dim(ctx, "rows", c(1L, 4L), 4L, "INT32"),
                                 tiledb_dim(ctx, "cols", c(1L, 4L), 4L, "INT32"))),
              attrs = c(tiledb_attr(ctx, "a", type = "INT32")), 
              cell_order = "COL_MAJOR",
              tile_order = "COL_MAJOR",
              sparse = FALSE)
schema 

}
