% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{tiledb_array_schema_dump}
\alias{tiledb_array_schema_dump}
\title{Print an Array Schema}
\usage{
tiledb_array_schema_dump(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{A TileDB Array URI}

\item{ctx}{An optional tiledb_ctx object}
}
\value{
Nothing is returned as the function is invoked for its site effect.
}
\description{
Prints the schema of the given array to standard output.
}
