% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\docType{class}
\name{tiledb_array_schema_evolution-class}
\alias{tiledb_array_schema_evolution-class}
\title{An S4 class for a TileDB ArraySchemaEvolution object}
\description{
An S4 class for a TileDB ArraySchemaEvolution object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

