% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution}
\alias{tiledb_array_schema_evolution}
\title{Creates a 'tiledb_array_schema_evolution' object}
\usage{
tiledb_array_schema_evolution(ctx = tiledb_get_context())
}
\arguments{
\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A 'array_schema_evolution' object
}
\description{
Creates a 'tiledb_array_schema_evolution' object
}
