% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_add_attribute}
\alias{tiledb_array_schema_evolution_add_attribute}
\title{Add an Attribute to a TileDB Array Schema Evolution object}
\usage{
tiledb_array_schema_evolution_add_attribute(object, attr)
}
\arguments{
\item{object}{A TileDB 'array_schema_evolution' object}

\item{attr}{A TileDB attribute}
}
\value{
The modified 'array_schema_evolution' object, invisibly
}
\description{
Add an Attribute to a TileDB Array Schema Evolution object
}
