% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_add_enumeration}
\alias{tiledb_array_schema_evolution_add_enumeration}
\title{Add an Enumeration to a TileDB Array Schema Evolution object}
\usage{
tiledb_array_schema_evolution_add_enumeration(
  object,
  name,
  enums,
  ordered = FALSE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{object}{A TileDB 'array_schema_evolution' object}

\item{name}{A character value with the name for the Enumeration}

\item{enums}{A character vector}

\item{ordered}{(optional) A boolean switch whether the enumeration is ordered}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
The modified 'array_schema_evolution' object, invisibly
}
\description{
Add an Enumeration to a TileDB Array Schema Evolution object
}
