% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_add_enumeration_empty}
\alias{tiledb_array_schema_evolution_add_enumeration_empty}
\title{Evolve an Array Schema by adding an empty Enumeration}
\usage{
tiledb_array_schema_evolution_add_enumeration_empty(
  ase,
  enum_name,
  type_str = "ASCII",
  cell_val_num = NA_integer_,
  ordered = FALSE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{ase}{An ArraySchemaEvolution object}

\item{enum_name}{A character value with the Enumeration name}

\item{type_str}{A character value with the TileDB type, defaults to \sQuote{ASCII}}

\item{cell_val_num}{An integer with number values per cell, defaults to \code{NA_integer_} to
flag the \code{NA} value use for character values}

\item{ordered}{A logical value indicating standard \code{factor} (when \code{FALSE}, the default)
or \code{ordered} (when \code{TRUE})}

\item{ctx}{Optional tiledb_ctx object}
}
\description{
Evolve an Array Schema by adding an empty Enumeration
}
