% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_array_evolve}
\alias{tiledb_array_schema_evolution_array_evolve}
\title{Evolve an Array Schema}
\usage{
tiledb_array_schema_evolution_array_evolve(object, uri)
}
\arguments{
\item{object}{A TileDB 'array_schema_evolution' object}

\item{uri}{A character variable with an URI}
}
\value{
The modified 'array_schema_evolution' object, invisibly
}
\description{
Evolve an Array Schema
}
