% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_drop_enumeration}
\alias{tiledb_array_schema_evolution_drop_enumeration}
\title{Drop an Enumeration given by name from a TileDB Array Schema Evolution object}
\usage{
tiledb_array_schema_evolution_drop_enumeration(object, attrname)
}
\arguments{
\item{object}{A TileDB 'array_schema_evolution' object}

\item{attrname}{A character variable with an attribute name}
}
\value{
The modified 'array_schema_evolution' object, invisibly
}
\description{
Drop an Enumeration given by name from a TileDB Array Schema Evolution object
}
