% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchemaEvolution.R
\name{tiledb_array_schema_evolution_expand_current_domain}
\alias{tiledb_array_schema_evolution_expand_current_domain}
\title{Expand an the Current Domain of an Array via Array Schema Evolution}
\usage{
tiledb_array_schema_evolution_expand_current_domain(ase, cd)
}
\arguments{
\item{ase}{An ArraySchemaEvolution object}

\item{cd}{A CurrentDomain object}
}
\value{
The modified ArraySchemaEvolution object
}
\description{
Expand an the Current Domain of an Array via Array Schema Evolution
}
