% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_get_allows_dups}
\alias{tiledb_array_schema_get_allows_dups}
\title{Returns logical value whether the array schema allows duplicate values or not.
This is only valid for sparse arrays.}
\usage{
tiledb_array_schema_get_allows_dups(x)
}
\arguments{
\item{x}{tiledb_array_schema}
}
\value{
the logical value
}
\description{
Returns logical value whether the array schema allows duplicate values or not.
This is only valid for sparse arrays.
}
