% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{capacity}
\alias{capacity}
\alias{capacity,tiledb_array_schema-method}
\alias{tiledb_array_schema_get_capacity}
\title{Retrieve schema capacity (for sparse fragments)}
\usage{
capacity(object)

\S4method{capacity}{tiledb_array_schema}(object)

tiledb_array_schema_get_capacity(object)
}
\arguments{
\item{object}{An \code{array_schema} object}
}
\value{
The tile capacity value
}
\description{
Returns the \code{tiledb_array} schema tile capacity for sparse fragments.
}
