% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_get_current_domain}
\alias{tiledb_array_schema_get_current_domain}
\title{Get the Current Domain of an Array Schema}
\usage{
tiledb_array_schema_get_current_domain(schema, ctx = tiledb_get_context())
}
\arguments{
\item{schema}{A TileDB Schema object}

\item{ctx}{Optional tiledb_ctx object}
}
\value{
An object of class \code{tiledb_current_domain}.
}
\description{
Note that \code{tiledb_current_domain} object may be empty.
}
