% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_set_allows_dups}
\alias{tiledb_array_schema_set_allows_dups}
\title{Sets toggle whether the array schema allows duplicate values or not.
This is only valid for sparse arrays.}
\usage{
tiledb_array_schema_set_allows_dups(x, allows_dups)
}
\arguments{
\item{x}{tiledb_array_schema}

\item{allows_dups}{logical value}
}
\value{
the logical value, invisibly
}
\description{
Sets toggle whether the array schema allows duplicate values or not.
This is only valid for sparse arrays.
}
