% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_set_coords_filter_list}
\alias{tiledb_array_schema_set_coords_filter_list}
\title{Set a Filter List for Coordinate of a TileDB Schema}
\usage{
tiledb_array_schema_set_coords_filter_list(sch, fl)
}
\arguments{
\item{sch}{A TileDB Schema object}

\item{fl}{A TileDB Filter List object}
}
\value{
The modified \code{tiledb_array_schema} object.
}
\description{
Set a Filter List for Coordinate of a TileDB Schema
}
