% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_set_current_domain}
\alias{tiledb_array_schema_set_current_domain}
\title{Set a Current Domain of an Array Schema}
\usage{
tiledb_array_schema_set_current_domain(schema, cd, ctx = tiledb_get_context())
}
\arguments{
\item{schema}{An Array Schema}

\item{cd}{An CurrendDomain object}

\item{ctx}{Optional tiledb_ctx object}
}
\value{
Nothing is returned from this function (but an error, should it occur is reported)
}
\description{
Set a Current Domain of an Array Schema
}
