% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_version}
\alias{tiledb_array_schema_version}
\title{Check the version of the array schema}
\usage{
tiledb_array_schema_version(object)
}
\arguments{
\item{object}{An \code{array_schema} object}
}
\value{
An integer value describing the internal schema format version
}
\description{
Returns the (internal) version of the \code{tiledb_array} schema
}
