% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_upgrade_version}
\alias{tiledb_array_upgrade_version}
\title{Upgrade an Array to the current TileDB Array Schema Format}
\usage{
tiledb_array_upgrade_version(array, config = NULL, ctx = tiledb_get_context())
}
\arguments{
\item{array}{A TileDB Array object}

\item{config}{A TileDB Configuration (optional, default NULL)}

\item{ctx}{A tiledb_ctx object (optional)}
}
\value{
Nothing is returned as the function is invoked for its side effect
}
\description{
Upgrade an Array to the current TileDB Array Schema Format
}
