% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowIO.R
\name{tiledb_arrow_array_ptr}
\alias{tiledb_arrow_array_ptr}
\alias{tiledb_arrow_schema_ptr}
\alias{tiledb_arrow_array_del}
\alias{tiledb_arrow_schema_del}
\title{Allocate (or Release) Arrow Array and Schema Pointers}
\usage{
tiledb_arrow_array_ptr()

tiledb_arrow_schema_ptr()

tiledb_arrow_array_del(ptr)

tiledb_arrow_schema_del(ptr)
}
\arguments{
\item{ptr}{A pointer object previously allocated with these functions}
}
\value{
The allocating functions return the requested pointer
}
\description{
These functions allocate (and free) appropriate pointer objects
for, respectively, Arrow array and schema objects.
}
