% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{class}
\name{tiledb_attr-class}
\alias{tiledb_attr-class}
\title{An S4 class for a TileDB attribute}
\description{
An S4 class for a TileDB attribute
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying implementation}
}}

