% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attr}
\alias{tiledb_attr}
\title{Contructs a \code{tiledb_attr} object}
\usage{
tiledb_attr(ctx, name = "", type = "FLOAT64",
  compressor = tiledb_compressor(), ncells = 1)
}
\arguments{
\item{name}{(default "") The dimension name / label string.}

\item{type}{(default "FLOAT64") The tiledb_attr TileDB datatype string}

\item{compressor}{(default tiledb_compressor("NOCOMPRESSION")) The tiledb_attr compressor}

\item{ncells}{(default 1) The number of cells}
}
\value{
\code{tiledb_dim} object
}
\description{
Contructs a \code{tiledb_attr} object
}
\examples{
ctx <- tiledb_ctx()
attr <- tiledb_attr(ctx, name = "a1", type = "INT32",
                    compressor = tiledb_compressor("GZIP", 5))
attr 
 
}
