% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_cell_size}
\alias{tiledb_attribute_get_cell_size}
\title{Get the TileDB Attribute cell size}
\usage{
tiledb_attribute_get_cell_size(attr)
}
\arguments{
\item{attr}{A TileDB Attribute object}
}
\value{
A numeric value with the cell size
}
\description{
Get the TileDB Attribute cell size
}
