% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_dictionary}
\alias{tiledb_attribute_get_dictionary}
\title{Get the TileDB Attribute Dictionary}
\usage{
tiledb_attribute_get_dictionary(attr)
}
\arguments{
\item{attr}{A TileDB Attribute object}
}
\value{
A character vector with the dictionary (of length zero if none)
}
\description{
Get the TileDB Attribute Dictionary
}
