% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_enumeration}
\alias{tiledb_attribute_get_enumeration}
\title{Get the TileDB Attribute Enumeration}
\usage{
tiledb_attribute_get_enumeration(attr, arr, ctx = tiledb_get_context())
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{arr}{A Tiledb Array object}
}
\value{
A character vector with the enumeration (of length zero if none)
}
\description{
Get the TileDB Attribute Enumeration
}
