% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_fill_value}
\alias{tiledb_attribute_get_fill_value}
\title{Get the fill value for a TileDB Attribute}
\usage{
tiledb_attribute_get_fill_value(attr)
}
\arguments{
\item{attr}{A TileDB Attribute object}
}
\value{
The fill value for the attribute
}
\description{
Get the fill value for a TileDB Attribute
}
