% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_nullable}
\alias{tiledb_attribute_get_nullable}
\title{Get the TileDB Attribute Nullable flag value}
\usage{
tiledb_attribute_get_nullable(attr)
}
\arguments{
\item{attr}{A TileDB Attribute object}
}
\value{
A boolean value with the \sQuote{Nullable} status
}
\description{
Get the TileDB Attribute Nullable flag value
}
