% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_has_enumeration}
\alias{tiledb_attribute_has_enumeration}
\title{Test if TileDB Attribute has an Enumeration}
\usage{
tiledb_attribute_has_enumeration(attr, ctx = tiledb_get_context())
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{ctx}{A Tiledb Context object (optional)}
}
\value{
A logical value indicating if the attribute has an enumeration
}
\description{
Test if TileDB Attribute has an Enumeration
}
