% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_is_ordered_enumeration_ptr}
\alias{tiledb_attribute_is_ordered_enumeration_ptr}
\title{Check if TileDB Attribute Enumeration is Ordered}
\usage{
tiledb_attribute_is_ordered_enumeration_ptr(
  attr,
  arrptr,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{attr}{A Tiledb Array object}

\item{arrptr}{A Tiledb Array object pointer}

\item{ctx}{A Tiledb Context object (optional)}
}
\value{
A character vector with the enumeration (of length zero if none)
}
\description{
Check if TileDB Attribute Enumeration is Ordered
}
