% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_is_variable_sized}
\alias{tiledb_attribute_is_variable_sized}
\title{Check whether TileDB Attribute is variable-sized}
\usage{
tiledb_attribute_is_variable_sized(attr)
}
\arguments{
\item{attr}{A TileDB Attribute object}
}
\value{
A boolean value indicating variable-size or not
}
\description{
Check whether TileDB Attribute is variable-sized
}
