% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{cell_val_num<-}
\alias{cell_val_num<-}
\alias{cell_val_num<-,tiledb_attr-method}
\alias{tiledb_attribute_set_cell_val_num}
\title{Set the number of scalar values per attribute cell}
\usage{
cell_val_num(x) <- value

\S4method{cell_val_num}{tiledb_attr}(x) <- value

tiledb_attribute_set_cell_val_num(x, value)
}
\arguments{
\item{x}{A TileDB Attribute object}

\item{value}{An integer value of number of cells}
}
\value{
The modified attribute is returned
}
\description{
Set the number of scalar values per attribute cell
}
