% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_set_dictionary}
\alias{tiledb_attribute_set_dictionary}
\title{Set a TileDB Attribute Dictionary}
\usage{
tiledb_attribute_set_dictionary(
  attr,
  vec,
  nullable = FALSE,
  ordered = FALSE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{vec}{A Character Vector with the dictionary values}
}
\value{
The modified TileDB Attribute object
}
\description{
Set a TileDB Attribute Dictionary
}
