% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_set_enumeration_name}
\alias{tiledb_attribute_set_enumeration_name}
\title{Set a TileDB Attribute Enumeration Name}
\usage{
tiledb_attribute_set_enumeration_name(
  attr,
  enum_name,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{enum_name}{A character value with the enumeration value}

\item{ctx}{A Tiledb Context object (optional)}
}
\value{
The modified TileDB Attribute object
}
\description{
Set a TileDB Attribute Enumeration Name
}
