% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_set_nullable}
\alias{tiledb_attribute_set_nullable}
\title{Set the TileDB Attribute Nullable flags}
\usage{
tiledb_attribute_set_nullable(attr, flag)
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{flag}{A boolean flag to turn \sQuote{Nullable} on or off}
}
\value{
Nothing is returned
}
\description{
Set the TileDB Attribute Nullable flags
}
