% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compressor.R
\name{tiledb_compressor}
\alias{tiledb_compressor}
\title{Constructs a \code{tiledb_compressor} object}
\usage{
tiledb_compressor(name = "NO_COMPRESSION", level = -1L)
}
\arguments{
\item{name}{(default "NO_COMPRESSION") TileDB compressor name string}

\item{level}{(default -1) compression level, -1 will fallback to the compression algorithm's default}
}
\value{
tiledb_compressor object
}
\description{
Available compressors:
\itemize{
\item "GZIP"
\item "ZSTD"
\item "LZ4"
\item "BLOSC_LZ"
\item "BLOSC_LZ4"
\item "BLOSC_LZ4HC"
\item "BLOSC_SNAPPY"
\item "BLOSC_ZSTD"
\item "RLE"
\item "BZIP2"
\item "DOUBLE_DELTA"
}
}
\details{
Valid compression levels vary depending on the compressor used,
consult the TileDB docs for more information.
}
\examples{
c <- tiledb_compressor("ZSTD", 3)
c

}
