% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compressor.R
\name{tiledb_compressor_level}
\alias{tiledb_compressor_level}
\title{Returns the compressor's compression level}
\usage{
tiledb_compressor_level(object)
}
\arguments{
\item{object}{tiledb_compressor}
}
\value{
integer level
}
\description{
Returns the compressor's compression level
}
\examples{
c <- tiledb_compressor("ZSTD", 3)
compressor_level(c)

}
