% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compressor.R
\name{tiledb_compressor_name}
\alias{tiledb_compressor_name}
\title{Returns the name of the compression library used}
\usage{
tiledb_compressor_name(object)
}
\arguments{
\item{object}{tiledb_compressor}
}
\value{
TileDB compression name string
}
\description{
Returns the name of the compression library used
}
\examples{
c <- tiledb_compressor("ZSTD", 3)
tiledb_compressor_name(c)

}
