% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\docType{class}
\name{tiledb_config-class}
\alias{tiledb_config-class}
\title{An S4 class for a TileDB configuration}
\description{
An S4 class for a TileDB configuration
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

