% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config_as_built_json}
\alias{tiledb_config_as_built_json}
\title{Return the 'AsBuilt' JSON string}
\usage{
tiledb_config_as_built_json()
}
\value{
The JSON string containing 'AsBuilt' information
}
\description{
Return the 'AsBuilt' JSON string
}
