% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config_as_built_show}
\alias{tiledb_config_as_built_show}
\title{Display the 'AsBuilt' JSON string}
\usage{
tiledb_config_as_built_show()
}
\value{
Nothing is returned but as a side-effect the 'AsBuilt' string is displayed
}
\description{
Display the 'AsBuilt' JSON string
}
