% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config_save}
\alias{tiledb_config_save}
\title{Save a TileDB Config object to a local text file}
\usage{
tiledb_config_save(config, path)
}
\arguments{
\item{config}{A \code{tiledb_config} object}

\item{path}{The path to config file to be created}
}
\value{
path to created config file
}
\description{
Save a TileDB Config object to a local text file
}
\examples{
\dontshow{
ctx <- tiledb_ctx(limitTileDBCores())
}
tmp <- tempfile()
cfg <- tiledb_config(c("sm.tile_cache_size" = "10"))
pth <- tiledb_config_save(cfg, tmp)

cat(readLines(pth), sep = "\n")

}
