% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config_unset}
\alias{tiledb_config_unset}
\title{Unset a TileDB Config parameter to its default value}
\usage{
tiledb_config_unset(config, param)
}
\arguments{
\item{config}{A TileDB Config object}

\item{param}{A character variable with the parameter name}
}
\value{
The modified TileDB Config object
}
\description{
Unset a TileDB Config parameter to its default value
}
