% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{tiledb_context_set_default_tags}
\alias{tiledb_context_set_default_tags}
\title{Sets default context tags}
\usage{
tiledb_context_set_default_tags(object)
}
\arguments{
\item{`tiledb_ctx`}{object}
}
\description{
Sets default context tags
}
