% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\docType{class}
\name{tiledb_ctx-class}
\alias{tiledb_ctx-class}
\title{An S4 class for a TileDB context}
\description{
An S4 class for a TileDB context
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

