% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{tiledb_ctx}
\alias{tiledb_ctx}
\title{Creates a \code{tiledb_ctx} object}
\usage{
tiledb_ctx(config = NULL)
}
\arguments{
\item{config}{(optonal) character vector of config parameter names, values}
}
\value{
\code{tiledb_ctx} object
}
\description{
Creates a \code{tiledb_ctx} object
}
\examples{
# default configuration
ctx <- tiledb_ctx()

# optionally set config parameters
ctx <- tiledb_ctx(c("sm.tile_cache_size" = "100")) 

}
