% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{tiledb_ctx_stats}
\alias{tiledb_ctx_stats}
\title{Return context statistics as a JSON string}
\usage{
tiledb_ctx_stats(object = tiledb_get_context())
}
\arguments{
\item{object}{\verb{A tiledb_ctx} object}
}
\value{
A JSON-formatted string with context statistics
}
\description{
Return context statistics as a JSON string
}
