% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\docType{class}
\name{tiledb_current_domain-class}
\alias{tiledb_current_domain-class}
\title{An S4 class for a TileDB CurrentDomain object}
\description{
An S4 class for a TileDB CurrentDomain object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying CurrentDomain object}
}}

