% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\name{tiledb_current_domain}
\alias{tiledb_current_domain}
\title{Creates a \code{tiledb_current_domain} object}
\usage{
tiledb_current_domain(ctx = tiledb_get_context())
}
\arguments{
\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
The \code{tiledb_current_domain} object
}
\description{
Creates a \code{tiledb_current_domain} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
if (tiledb_version(TRUE) >= "2.25.0") {
   cd <-tiledb_current_domain()
}

}
