% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\name{tiledb_current_domain_get_ndrectangle}
\alias{tiledb_current_domain_get_ndrectangle}
\title{Get a \code{tiledb_ndrectangle} from a \code{tiledb_current_domain} object}
\usage{
tiledb_current_domain_get_ndrectangle(cd)
}
\arguments{
\item{cd}{A TileDB CurrentDomain object}
}
\value{
The corresponding TileDB NDRectangle object
}
\description{
Get a \code{tiledb_ndrectangle} from a \code{tiledb_current_domain} object
}
