% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\name{tiledb_current_domain_get_type}
\alias{tiledb_current_domain_get_type}
\title{Get \code{tiledb_current_domain} data type as string}
\usage{
tiledb_current_domain_get_type(cd)
}
\arguments{
\item{cd}{A TileDB CurrentDomain object}
}
\value{
The datatype (as string) of the \code{tiledb_current_domain} object
}
\description{
Get \code{tiledb_current_domain} data type as string
}
