% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\name{tiledb_current_domain_is_empty}
\alias{tiledb_current_domain_is_empty}
\title{Test \code{tiledb_current_domain} object for being empty}
\usage{
tiledb_current_domain_is_empty(cd)
}
\arguments{
\item{cd}{A TileDB CurrentDomain object}
}
\value{
A boolean indicating whether the object is empty or not
}
\description{
Test \code{tiledb_current_domain} object for being empty
}
