% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurrentDomain.R
\name{tiledb_current_domain_set_ndrectangle}
\alias{tiledb_current_domain_set_ndrectangle}
\title{Set a \code{tiledb_ndrectangle} in a \code{tiledb_current_domain} object}
\usage{
tiledb_current_domain_set_ndrectangle(cd, ndr)
}
\arguments{
\item{cd}{A TileDB CurrentDomain object}

\item{ndr}{A TileDB NDRectangle object}
}
\value{
The modifiled TileDB CurrendDomain object
}
\description{
Set a \code{tiledb_ndrectangle} in a \code{tiledb_current_domain} object
}
