% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tiledb_datatype_R_type}
\alias{tiledb_datatype_R_type}
\title{Map from TileDB type to R datatype}
\usage{
tiledb_datatype_R_type(datatype)
}
\arguments{
\item{datatype}{A string describing one TileDB datatype}
}
\value{
A string describing the closest match for an R datatype
}
\description{
This function maps from the TileDB types to the (fewer) key datatypes in R. This
can be lossy as TileDB integers range from (signed and unsigned) 8 to 64 bit whereas
R only has (signed) 32 bit values. Similarly, R only has 64 bit doubles whereas TileDB
has 32 and 64 bit floating point types. TileDB also has more character encodings, and the
full range of (NumPy) date and time types.
}
