% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_delete_metadata}
\alias{tiledb_delete_metadata}
\title{Delete a TileDB Array Metadata object given by key}
\usage{
tiledb_delete_metadata(arr, key)
}
\arguments{
\item{arr}{A TileDB Array object}

\item{key}{A character value describing a metadata key}
}
\value{
A boolean indicating success
}
\description{
Delete a TileDB Array Metadata object given by key
}
