% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\docType{class}
\name{tiledb_dense-class}
\alias{tiledb_dense-class}
\title{An S4 class for a TileDB dense array}
\description{
An S4 class for a TileDB dense array
}
\section{Slots}{

\describe{
\item{\code{ctx}}{A TileDB context object}

\item{\code{uri}}{A character despription}

\item{\code{as.data.frame}}{A logical value}

\item{\code{attrs}}{A character vector}

\item{\code{extended}}{A logical value}

\item{\code{ptr}}{External pointer to the underlying implementation}
}}

