% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{tiledb_dense}
\alias{tiledb_dense}
\title{Constructs a tiledb_dense object backed by a persisted tiledb array uri}
\usage{
tiledb_dense(ctx, uri, query_type = c("READ", "WRITE"),
  as.data.frame = FALSE)
}
\arguments{
\item{ctx}{tiledb_ctx}

\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}
}
\value{
tiledb_dense array object
}
\description{
Constructs a tiledb_dense object backed by a persisted tiledb array uri
}
