% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\docType{class}
\name{tiledb_dim-class}
\alias{tiledb_dim-class}
\title{An S4 class for a TileDB dimension object}
\description{
An S4 class for a TileDB dimension object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

