% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{cell_val_num,tiledb_dim-method}
\alias{cell_val_num,tiledb_dim-method}
\alias{tiledb_dim_get_cell_val_num}
\title{Return the number of scalar values per dimension cell}
\usage{
\S4method{cell_val_num}{tiledb_dim}(object)

tiledb_dim_get_cell_val_num(object)
}
\arguments{
\item{object}{\code{tiledb_dim} object}
}
\value{
integer number of cells
}
\description{
Return the number of scalar values per dimension cell
}
