% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\docType{class}
\name{tiledb_domain-class}
\alias{tiledb_domain-class}
\title{An S4 class for a TileDB domain}
\description{
An S4 class for a TileDB domain
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying implementation}
}}

