% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain}
\alias{tiledb_domain}
\title{Constructs a \code{tiledb_domain} object
All dimension's must be of the same TileDB type.}
\usage{
tiledb_domain(ctx, dims)
}
\arguments{
\item{ctx}{tiledb_ctx}

\item{dims}{list() of tiledb_dim objects}
}
\value{
tiledb_domain
}
\description{
Constructs a \code{tiledb_domain} object
All dimension's must be of the same TileDB type.
}
\examples{
ctx <- tiledb_ctx()
dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(1L, 100L), type = "INT32"),
                                   tiledb_dim(ctx, "d2", c(1L, 50L), type = "INT32")))
 
}
