% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain_get_dimension_from_index}
\alias{tiledb_domain_get_dimension_from_index}
\title{Returns a Dimension indicated by index for the given TileDB Domain}
\usage{
tiledb_domain_get_dimension_from_index(domain, idx)
}
\arguments{
\item{domain}{TileDB Domain object}

\item{idx}{Integer index of the selected dimension}
}
\value{
TileDB Dimension object
}
\description{
Returns a Dimension indicated by index for the given TileDB Domain
}
