% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain_get_dimension_from_name}
\alias{tiledb_domain_get_dimension_from_name}
\title{Returns a Dimension indicated by name for the given TileDB Domain}
\usage{
tiledb_domain_get_dimension_from_name(domain, name)
}
\arguments{
\item{domain}{TileDB Domain object}

\item{name}{A character variable with a dimension name}
}
\value{
TileDB Dimension object
}
\description{
Returns a Dimension indicated by name for the given TileDB Domain
}
