% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain_has_dimension}
\alias{tiledb_domain_has_dimension}
\title{Check a domain for a given dimension name}
\usage{
tiledb_domain_has_dimension(domain, name)
}
\arguments{
\item{domain}{A domain of a TileDB Array schema}

\item{name}{A character variable with a dimension name}
}
\value{
A boolean value indicating if the dimension exists in the domain
}
\description{
Check a domain for a given dimension name
}
