% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain_simple}
\alias{tiledb_domain_simple}
\title{Simplified constructor for tiledb domain}
\usage{
tiledb_domain_simple(dims)
}
\arguments{
\item{dims}{integer or numeric vector, one value per dimension}
}
\value{
\code{tiledb_domain}, which contains a list of \code{tiledb_dim}
}
\description{
Takes the dimensiosn of the array. Dimensions will be anonymous,
and have an INT32 domain from 1 to n for each dimension.
}
\examples{
tiledb_domain_simple(c(2,3))
tiledb_domain_simple(c(2L,3L))
}
