% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Error.R
\name{tiledb_error_message}
\alias{tiledb_error_message}
\title{Return the error message for a given context}
\usage{
tiledb_error_message(ctx = tiledb_get_context())
}
\arguments{
\item{ctx}{\verb{A tiledb_ctx} object}
}
\value{
A character variable with the error message
}
\description{
Note that this function requires an actual error to have occurred.
}
