% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_buffer_export}
\alias{tiledb_filestore_buffer_export}
\title{Export from a TileDB Filestore to a character variable}
\usage{
tiledb_filestore_buffer_export(
  filestore_uri,
  offset,
  bytes,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{filestore_uri}{Character with an TileDB Array Schema URI}

\item{offset}{(optional) Numeric variable with offset from beginning, default is zero}

\item{bytes}{(optional) Numeric variable with number of bytes to read, default is zero}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A character variable containing the filestore content (subject to offset and
bytes) is returned
}
\description{
Export from a TileDB Filestore to a character variable
}
