% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_buffer_import}
\alias{tiledb_filestore_buffer_import}
\title{Import size bytes from a string into a TileDB Filestore}
\usage{
tiledb_filestore_buffer_import(
  filestore_uri,
  buf,
  bytes,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{filestore_uri}{Character with an TileDB Array Schema URI}

\item{buf}{Character variable with content to be imported}

\item{bytes}{Number of bytes to be import, defaults to length of \code{buf}}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A boolean is returned to indicate successful completion
}
\description{
Import size bytes from a string into a TileDB Filestore
}
