% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_schema_create}
\alias{tiledb_filestore_schema_create}
\title{Create an array schema from a given URI}
\usage{
tiledb_filestore_schema_create(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{Character with an TileDB Array URI, if missing or NULL a default schema
is returned}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
An ArraySchema object
}
\description{
Create an array schema from a given URI
}
