% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_size}
\alias{tiledb_filestore_size}
\title{Return (uncompressed) TileDB Filestore size}
\usage{
tiledb_filestore_size(filestore_uri, ctx = tiledb_get_context())
}
\arguments{
\item{filestore_uri}{Character with an TileDB Array Schema URI}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A numeric with the size is returned
}
\description{
Return (uncompressed) TileDB Filestore size
}
