% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_uri_export}
\alias{tiledb_filestore_uri_export}
\title{Export a file from a TileDB Filestore}
\usage{
tiledb_filestore_uri_export(
  file_uri,
  filestore_uri,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{file_uri}{Character with a file URI}

\item{filestore_uri}{Character with an TileDB Array Schema URI}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A boolean is returned to indicate successful completion
}
\description{
Export a file from a TileDB Filestore
}
