% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filestore.R
\name{tiledb_filestore_uri_import}
\alias{tiledb_filestore_uri_import}
\title{Import a file into a TileDB Filestore}
\usage{
tiledb_filestore_uri_import(
  filestore_uri,
  file_uri,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{filestore_uri}{Character with an TileDB Array Schema URI}

\item{file_uri}{Character with a file URI}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A boolean is returned to indicate successful completion
}
\description{
Import a file into a TileDB Filestore
}
