% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\docType{class}
\name{tiledb_filter-class}
\alias{tiledb_filter-class}
\title{An S4 class for a TileDB filter}
\description{
An S4 class for a TileDB filter
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying implementation}
}}

