% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{tiledb_filter}
\alias{tiledb_filter}
\title{Constructs a \code{tiledb_filter} object}
\usage{
tiledb_filter(ctx = tiledb::ctx, name = "NONE")
}
\arguments{
\item{ctx}{tiledb_ctx object}

\item{name}{(default "NONE") TileDB filter name string}
}
\value{
tiledb_filter object
}
\description{
Available filters:
\itemize{
\item "NONE"
\item "GZIP"
\item "ZSTD"
\item "LZ4"
\item "RLE"
\item "BZIP2"
\item "DOUBLE_DELTA"
\item "BIT_WIDTH_REDUCTION"
\item "BITSHUFFLE"
\item "BYTESHUFFLE"
\item "POSITIVE_DELTA"
}
}
\details{
Valid compression options vary depending on the filter used,
consult the TileDB docs for more information.
}
\examples{
c <- tiledb_filter(ctx, "ZSTD")
c

}
