% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\docType{class}
\name{tiledb_filter_list-class}
\alias{tiledb_filter_list-class}
\title{An S4 class for a TileDB filter list}
\description{
An S4 class for a TileDB filter list
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

