% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\name{max_chunk_size}
\alias{max_chunk_size}
\alias{max_chunk_size,tiledb_filter_list-method}
\alias{tiledb_filter_list_get_max_chunk_size}
\title{Returns the filter_list's max_chunk_size}
\usage{
max_chunk_size(object)

\S4method{max_chunk_size}{tiledb_filter_list}(object)

tiledb_filter_list_get_max_chunk_size(object)
}
\arguments{
\item{object}{tiledb_filter_list}
}
\value{
integer max_chunk_size
}
\description{
Returns the filter_list's max_chunk_size
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
flt <- tiledb_filter("ZSTD")
tiledb_filter_set_option(flt, "COMPRESSION_LEVEL", 5)
filter_list <- tiledb_filter_list(c(flt))
max_chunk_size(filter_list)

}
