% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\docType{class}
\name{tiledb_fragment_info-class}
\alias{tiledb_fragment_info-class}
\title{An S4 class for a TileDB fragment info object}
\description{
An S4 class for a TileDB fragment info object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

