% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info}
\alias{tiledb_fragment_info}
\title{Constructs a \code{tiledb_fragment_info} object}
\usage{
tiledb_fragment_info(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{an character variable with the URI of the
array for which fragment info is request}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
tiledb_fragment_info object
}
\description{
Constructs a \code{tiledb_fragment_info} object
}
