% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_dense}
\alias{tiledb_fragment_info_dense}
\title{Return if a fragment info index is dense}
\usage{
tiledb_fragment_info_dense(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A logical value indicating if the fragment is dense
}
\description{
Return if a fragment info index is dense
}
