% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_cell_num}
\alias{tiledb_fragment_info_get_cell_num}
\title{Return a fragment info number of cells for a given fragment index}
\usage{
tiledb_fragment_info_get_cell_num(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A numeric value with the number of cells
}
\description{
Return a fragment info number of cells for a given fragment index
}
