% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_non_empty_domain_index}
\alias{tiledb_fragment_info_get_non_empty_domain_index}
\title{Return a fragment info non-empty domain from index}
\usage{
tiledb_fragment_info_get_non_empty_domain_index(object, fid, did, typestr)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}

\item{did}{A domain index}

\item{typestr}{An optional character variable describing the data type which will
be accessed from the schema if missinh}
}
\value{
A TileDB Domain object
}
\description{
TODO: Rework with type information
}
