% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_non_empty_domain_var_index}
\alias{tiledb_fragment_info_get_non_empty_domain_var_index}
\title{Return a fragment info non-empty domain variable from index}
\usage{
tiledb_fragment_info_get_non_empty_domain_var_index(object, fid, did)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}

\item{did}{A domain index}
}
\value{
A character vector with two elements
}
\description{
Return a fragment info non-empty domain variable from index
}
