% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_num}
\alias{tiledb_fragment_info_get_num}
\title{Return a fragment info number of fragments}
\usage{
tiledb_fragment_info_get_num(object, fid)

tiledb_fragment_info_get_num(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A numeric variable with the number of fragments

A numeric variable with the number of fragments
}
\description{
Return a fragment info number of fragments

Return a fragment info fragment size for a given fragment index
}
