% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_timestamp_range}
\alias{tiledb_fragment_info_get_timestamp_range}
\title{Return a fragment info timestamp range for a given fragment index}
\usage{
tiledb_fragment_info_get_timestamp_range(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A Datetime vector with two elements for the range
}
\description{
Return a fragment info timestamp range for a given fragment index
}
