% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_to_vacuum_num}
\alias{tiledb_fragment_info_get_to_vacuum_num}
\title{Return the number of fragment info elements to be vacuumed}
\usage{
tiledb_fragment_info_get_to_vacuum_num(object)
}
\arguments{
\item{object}{A TileDB fragment info object}
}
\value{
A numeric value with the number of to be vacuumed fragments
}
\description{
Return the number of fragment info elements to be vacuumed
}
