% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_to_vacuum_uri}
\alias{tiledb_fragment_info_get_to_vacuum_uri}
\title{Return fragment info URI of the to be vacuumed index}
\usage{
tiledb_fragment_info_get_to_vacuum_uri(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A character variable with the URI of the be vacuumed index
}
\description{
Return fragment info URI of the to be vacuumed index
}
