% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_unconsolidated_metadata_num}
\alias{tiledb_fragment_info_get_unconsolidated_metadata_num}
\title{Return fragment info number of unconsolidated metadata}
\usage{
tiledb_fragment_info_get_unconsolidated_metadata_num(object)
}
\arguments{
\item{object}{A TileDB fragment info object}
}
\value{
A numeric value with the number of unconsolidated metadata
}
\description{
Return fragment info number of unconsolidated metadata
}
