% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_version}
\alias{tiledb_fragment_info_get_version}
\title{Return a fragment info version for a given fragment index}
\usage{
tiledb_fragment_info_get_version(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A integer value value with the version
}
\description{
Return a fragment info version for a given fragment index
}
