% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_has_consolidated_metadata}
\alias{tiledb_fragment_info_has_consolidated_metadata}
\title{Return if a fragment info index has consolidated metadata}
\usage{
tiledb_fragment_info_has_consolidated_metadata(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A logical value indicating consolidated metadata
}
\description{
Return if a fragment info index has consolidated metadata
}
