% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_sparse}
\alias{tiledb_fragment_info_sparse}
\title{Return if a fragment info index is sparse}
\usage{
tiledb_fragment_info_sparse(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A logical value indicating if the fragment is sparse
}
\description{
Return if a fragment info index is sparse
}
