% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_uri}
\alias{tiledb_fragment_info_uri}
\title{Return a fragment info URI given its index}
\usage{
tiledb_fragment_info_uri(object, fid)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}
}
\value{
A character variable with URI
}
\description{
Return a fragment info URI given its index
}
