% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_get_all_metadata}
\alias{tiledb_get_all_metadata}
\title{Return a TileDB Array Metadata object given by key}
\usage{
tiledb_get_all_metadata(arr)
}
\arguments{
\item{arr}{A TileDB Array object, or a character URI describing one}
}
\value{
A object stored in the Metadata under the given key
}
\description{
Return a TileDB Array Metadata object given by key
}
