% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{tiledb_get_context}
\alias{tiledb_get_context}
\title{Retrieve a TileDB context object from the package cache}
\usage{
tiledb_get_context()
}
\value{
A TileDB context object
}
\description{
Retrieve a TileDB context object from the package cache
}
