% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_get_metadata}
\alias{tiledb_get_metadata}
\title{Return a TileDB Array Metadata object given by key}
\usage{
tiledb_get_metadata(arr, key)
}
\arguments{
\item{arr}{A TileDB Array object, or a character URI describing one}

\item{key}{A character value describing a metadata key}
}
\value{
A object stored in the Metadata under the given key,
or \sQuote{NULL} if none found.
}
\description{
Return a TileDB Array Metadata object given by key
}
