% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_get_query_status}
\alias{tiledb_get_query_status}
\title{Retrieve the cached status of the last finalized query}
\usage{
tiledb_get_query_status()
}
\value{
The status of the last query
}
\description{
This function accesses the status of the last query without
requiring the query object.
}
