% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_get_vfs}
\alias{tiledb_get_vfs}
\title{Retrieve a TileDB VFS object from the package environment and cache}
\usage{
tiledb_get_vfs()
}
\value{
A TileDB VFS object
}
\description{
Retrieve a TileDB VFS object from the package environment and cache
}
