% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\docType{class}
\name{tiledb_group-class}
\alias{tiledb_group-class}
\title{An S4 class for a TileDB Group object}
\description{
An S4 class for a TileDB Group object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

