% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group}
\alias{tiledb_group}
\title{Creates a 'tiledb_group' object}
\usage{
tiledb_group(uri, type = c("READ", "WRITE"), ctx = tiledb_get_context())
}
\arguments{
\item{uri}{Character variable with the URI of the new group object}

\item{type}{Character variable with the query type value: one of \dQuote{READ}
or \dQuote{WRITE}}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A 'group' object
}
\description{
Creates a 'tiledb_group' object
}
