% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_close}
\alias{tiledb_group_close}
\title{Close a TileDB Group}
\usage{
tiledb_group_close(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
The TileDB Group object but closed for reading or writing
}
\description{
Close a TileDB Group
}
