% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{tiledb_group_create}
\alias{tiledb_group_create}
\title{Creates a TileDB group object at given uri path}
\usage{
tiledb_group_create(ctx = tiledb:::ctx, uri)
}
\arguments{
\item{ctx}{tiledb_ctx}

\item{uri}{path which to create group}
}
\value{
uri of created group
}
\description{
Creates a TileDB group object at given uri path
}
\examples{
pth <- tempdir()
ctx <- tiledb_ctx()
tiledb_group_create(ctx, pth)
tiledb_object_type(ctx, pth)

}
