% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_delete}
\alias{tiledb_group_delete}
\title{Deletes all written data from a 'tiledb_group' object}
\usage{
tiledb_group_delete(grp, uri, recursive = FALSE)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{uri}{Character variable with the URI of the group item to be deleted}

\item{recursive}{A logical value indicating whether all data iniside the
group is to be delet}
}
\value{
Nothing is returned, the function is invoked for the side-effect of
group data removal.
}
\description{
The group must be opened in \sQuote{MODIFY_EXCLUSIVE} mode, otherwise the function
will error out.
}
