% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_get_all_metadata}
\alias{tiledb_group_get_all_metadata}
\title{Return all Metadata from a TileDB Group}
\usage{
tiledb_group_get_all_metadata(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
A named List with all Metadata objects index
}
\description{
Return all Metadata from a TileDB Group
}
