% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_get_config}
\alias{tiledb_group_get_config}
\title{Get a TileDB Config from a TileDB Group}
\usage{
tiledb_group_get_config(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
The TileDB Config object of the TileDB Group object
}
\description{
Get a TileDB Config from a TileDB Group
}
