% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_get_metadata}
\alias{tiledb_group_get_metadata}
\title{Accesses Metadata from a TileDB Group}
\usage{
tiledb_group_get_metadata(grp, key)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{key}{A character value with the key of the metadata object to be retrieved}
}
\value{
The requested object, or NULL is not found
}
\description{
Accesses Metadata from a TileDB Group
}
