% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_get_metadata_from_index}
\alias{tiledb_group_get_metadata_from_index}
\title{Accesses Metadata by Index from a TileDB Group}
\usage{
tiledb_group_get_metadata_from_index(grp, idx)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{idx}{A numeric value with the index of the metadata object to be retrieved}
}
\value{
The requested object, or NULL is not found
}
\description{
Accesses Metadata by Index from a TileDB Group
}
