% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_has_metadata}
\alias{tiledb_group_has_metadata}
\title{Checks for Metadata in a TileDB Group}
\usage{
tiledb_group_has_metadata(grp, key)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{key}{A character value with they index under which the data will be written}
}
\value{
A boolean value indicating with the object is present
}
\description{
Checks for Metadata in a TileDB Group
}
