% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_is_open}
\alias{tiledb_group_is_open}
\title{Test if TileDB Group is open}
\usage{
tiledb_group_is_open(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
A boolean indicating whether the TileDB Group object is open
}
\description{
Test if TileDB Group is open
}
