% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_is_relative}
\alias{tiledb_group_is_relative}
\title{Test if a Named Group is Using a Relative URI}
\usage{
tiledb_group_is_relative(grp, name)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{name}{A character value with a group name}
}
\value{
A boolean indicating whether the group uses a relative URI or not
}
\description{
Test if a Named Group is Using a Relative URI
}
