% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_member}
\alias{tiledb_group_member}
\title{Get a Member (Description) by Index from TileDB Group}
\usage{
tiledb_group_member(grp, idx)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{idx}{A numeric value with the index of the metadata object to be retrieved}
}
\value{
A character vector with two elements: the member type, and its uri.
}
\description{
This function returns a two-element character vector with the member object translated to
character, and the uri.
}
