% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_member_count}
\alias{tiledb_group_member_count}
\title{Get Member Count from TileDB Group}
\usage{
tiledb_group_member_count(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
The Count of Members in the TileDB Group object
}
\description{
Get Member Count from TileDB Group
}
