% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_metadata_num}
\alias{tiledb_group_metadata_num}
\title{Returns Number of Metadata Objects a TileDB Group}
\usage{
tiledb_group_metadata_num(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
A numeric value with the number of metadata objects
}
\description{
Returns Number of Metadata Objects a TileDB Group
}
