% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_open}
\alias{tiledb_group_open}
\title{Open a TileDB Group}
\usage{
tiledb_group_open(grp, type = c("READ", "WRITE"))
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{type}{A character value that must be either \sQuote{READ} or \sQuote{WRITE}}
}
\value{
The TileDB Group object but opened for reading or writing
}
\description{
Open a TileDB Group
}
