% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_put_metadata}
\alias{tiledb_group_put_metadata}
\title{Write Metadata to a TileDB Group}
\usage{
tiledb_group_put_metadata(grp, key, val)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{key}{A character value with they index under which the data will be written}

\item{val}{An R object (numeric, int, or char vector) that will be stored}
}
\value{
On success boolean \sQuote{TRUE} is returned
}
\description{
Write Metadata to a TileDB Group
}
