% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_query_type}
\alias{tiledb_group_query_type}
\title{Return a TileDB Group query type}
\usage{
tiledb_group_query_type(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
A character value with the query type i.e. one of \dQuote{READ} or \dQuote{WRITE}.
}
\description{
Return a TileDB Group query type
}
