% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_remove_member}
\alias{tiledb_group_remove_member}
\title{Remove Member from TileDB Group}
\usage{
tiledb_group_remove_member(grp, uri)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{uri}{A character value with a the URI of the member to be removed, or (if added
with a name) the name of the member}
}
\value{
The TileDB Group object, invisibly
}
\description{
Remove Member from TileDB Group
}
