% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_set_config}
\alias{tiledb_group_set_config}
\title{Set a TileDB Config for a TileDB Group}
\usage{
tiledb_group_set_config(grp, cfg)

tiledb_group_set_config(grp, cfg)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}

\item{cfg}{A TileDB Config object}
}
\value{
The TileDB Group object with added Config

The TileDB Config object of the TileDB Group object
}
\description{
Set a TileDB Config for a TileDB Group

Get a TileDB Config from a TileDB Group
}
