% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_uri}
\alias{tiledb_group_uri}
\title{Return a TileDB Group URI}
\usage{
tiledb_group_uri(grp)
}
\arguments{
\item{grp}{A TileDB Group object as for example returned by \code{tiledb_group()}}
}
\value{
A character value with the URI
}
\description{
Return a TileDB Group URI
}
