% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_has_metadata}
\alias{tiledb_has_metadata}
\title{Test if TileDB Array has Metadata}
\usage{
tiledb_has_metadata(arr, key)
}
\arguments{
\item{arr}{A TileDB Array object}

\item{key}{A character value describing a metadata key}
}
\value{
A logical value indicating if the given key exists in the
metdata of the given array
}
\description{
Test if TileDB Array has Metadata
}
