% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\docType{methods}
\name{tiledb_ndim,tiledb_array_schema-method}
\alias{tiledb_ndim,tiledb_array_schema-method}
\title{Return the number of dimensions associated with the \code{tieldb_array_schema}}
\usage{
\S4method{tiledb_ndim}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{tiledb_array_schema}
}
\value{
integer number of dimensions
}
\description{
Return the number of dimensions associated with the \code{tieldb_array_schema}
}
\examples{
ctx <- tiledb_ctx()
dom <- tiledb_domain(ctx, dims = c(tiledb_dim(ctx, "d1", c(1L, 10L), type = "INT32")))
sch <- tiledb_array_schema(ctx, dom, attrs = c(tiledb_attr(ctx, "a1", type = "INT32"),
                                               tiledb_attr(ctx, "a2", type = "FLOAT64")))
tiledb_ndim(sch)

}
