% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\docType{methods}
\name{tiledb_ndim,tiledb_dim-method}
\alias{tiledb_ndim,tiledb_dim-method}
\title{Returns the number of dimensions for a tiledb domain object}
\usage{
\S4method{tiledb_ndim}{tiledb_dim}(object)
}
\arguments{
\item{`tiledb_ndim`}{object}
}
\value{
1L
}
\description{
Returns the number of dimensions for a tiledb domain object
}
\examples{
d1 <- tiledb_dim("d1", c(1L, 10L), 10L)
tiledb_ndim(d1)

}
