% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDRectangle.R
\docType{class}
\name{tiledb_ndrectangle-class}
\alias{tiledb_ndrectangle-class}
\title{An S4 class for a TileDB NDRectangle object}
\description{
An S4 class for a TileDB NDRectangle object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying NDRectangle object}
}}

