% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDRectangle.R
\name{tiledb_ndrectangle}
\alias{tiledb_ndrectangle}
\title{Creates a \code{tiledb_ndrectangle} object}
\usage{
tiledb_ndrectangle(dom, ctx = tiledb_get_context())
}
\arguments{
\item{ctx}{(optional) A TileDB Ctx object}

\item{domain}{A TileDB Domain object for which the NDRectangle object is created}
}
\value{
The \code{tiledb_ndrectangle} object
}
\description{
Creates a \code{tiledb_ndrectangle} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <-tiledb_domain(dim = tiledb_dim("d1", c(1L, 100L), type = "INT32"))
ndr <- tiledb_ndrectangle(dom)

}
