% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDRectangle.R
\name{tiledb_ndrectangle_dim_num}
\alias{tiledb_ndrectangle_dim_num}
\title{Get the number of dimensions for \code{tiledb_ndrectangle} object}
\usage{
tiledb_ndrectangle_dim_num(ndr)
}
\arguments{
\item{ndr}{A TileDB NDRectangle object}
}
\value{
The number of dimentiones for the \code{tiledb_ndrectangle}
}
\description{
Get the number of dimensions for \code{tiledb_ndrectangle} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
if (tiledb_version(TRUE) >= "2.26.0") {
   dom <- tiledb_domain(dim = tiledb_dim("d1", c(1L, 100L), type = "INT32"))
   ndr <- tiledb_ndrectangle(dom)
   tiledb_ndrectangle_dim_num(ndr)
}
}
