% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDRectangle.R
\name{tiledb_ndrectangle_set_range}
\alias{tiledb_ndrectangle_set_range}
\title{Set a range on a \code{tiledb_ndrectangle} object}
\usage{
tiledb_ndrectangle_set_range(ndr, dimname, start, end)
}
\arguments{
\item{ndr}{A TileDB NDRectangle object}

\item{dimname}{A character variable with the dimension for which to set a range}

\item{start}{The lower end of the range to be set}

\item{end}{The upper end of the range to be set}
}
\value{
The modified \code{tiledb_ndrectangle} object

Start and end values have to be of the same data type as the type of the selected
dimension. The set of allowed type includes the different integer types as well as
string dimensions.
}
\description{
Set a range on a \code{tiledb_ndrectangle} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <-tiledb_domain(dim = tiledb_dim("d1", c(1L, 100L), type = "INT32"))
ndr <- tiledb_ndrectangle(dom)
ndr <- tiledb_ndrectangle_set_range(ndr, "d1", 50, 500)

}
