% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_num_metadata}
\alias{tiledb_num_metadata}
\title{Return count of TileDB Array Metadata objects}
\usage{
tiledb_num_metadata(arr)
}
\arguments{
\item{arr}{A TileDB Array object, or a character URI describing one}
}
\value{
A integer variable with the number of Metadata objects
}
\description{
Return count of TileDB Array Metadata objects
}
