% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{tiledb_object_mv}
\alias{tiledb_object_mv}
\title{Move a TileDB resource to new uri path}
\usage{
tiledb_object_mv(old_uri, new_uri, ctx = tiledb_get_context())
}
\arguments{
\item{old_uri}{old uri of existing tiledb resource}

\item{new_uri}{new uri to move tiledb resource}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
new uri of moved tiledb resource
}
\description{
Raises an error if either uri is invalid, or the old uri resource is not a tiledb object
}
