% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{tiledb_object_rm}
\alias{tiledb_object_rm}
\title{Removes a TileDB resource}
\usage{
tiledb_object_rm(uri, ctx = tiledb:::ctx)
}
\arguments{
\item{uri}{path which to create group}
}
\value{
uri of removed TileDB resource
}
\description{
Raises an error if the uri is invalid, or the uri resource is not a tiledb object
}
