% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{tiledb_object_type}
\alias{tiledb_object_type}
\title{Return the TileDB object type string of a TileDB resource}
\usage{
tiledb_object_type(ctx, uri)
}
\arguments{
\item{ctx}{tiledb_ctx}

\item{uri}{path to TileDB resource}
}
\value{
TileDB object type string
}
\description{
Object types:
\itemize{
\item \code{"ARRAY"}, dense or sparse TileDB array
\item \code{"KEY_VALUE"}, TileDB kv array
\item \code{"GROUP"}, TileDB group
\item `"INVALID"``, not a TileDB resource
}
}
