% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\docType{class}
\name{tiledb_profile-class}
\alias{tiledb_profile-class}
\title{An S4 class for a TileDB Profile object}
\description{
An S4 class for a TileDB Profile object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

