% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile}
\alias{tiledb_profile}
\title{Create a 'tiledb_profile' object}
\usage{
tiledb_profile(name = NULL, dir = NULL)
}
\arguments{
\item{name}{(optional) Name for the profile}

\item{dir}{(optional) Directory to create the profile in}
}
\value{
A new 'tiledb_profile' object
}
\description{
Create a 'tiledb_profile' object
}
