% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_dir}
\alias{tiledb_profile_dir}
\title{Get the directory of a 'tiledb_profile' object}
\usage{
tiledb_profile_dir(profile)
}
\value{
The directory of the 'tiledb_profile' object
}
\description{
Get the directory of a 'tiledb_profile' object
}
