% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_get_param}
\alias{tiledb_profile_get_param}
\title{Get the value of a parameter set on the 'tiledb_profile' object}
\usage{
tiledb_profile_get_param(profile, param)
}
\arguments{
\item{profile}{A TileDB profile object}

\item{param}{The key for the parameter to fetch}
}
\value{
The value of the requested parameter or NULL if no such parameter exists
}
\description{
Get the value of a parameter set on the 'tiledb_profile' object
}
