% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_load}
\alias{tiledb_profile_load}
\title{Load a saved 'tiledb_profile' object}
\usage{
tiledb_profile_load(name = NULL, dir = NULL)
}
\arguments{
\item{name}{(optional) Name of the profile to load}

\item{dir}{(optional) Directory where the profile to load is saved}
}
\value{
The loaded 'tiledb_profile' object
}
\description{
Load a saved 'tiledb_profile' object
}
