% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_name}
\alias{tiledb_profile_name}
\title{Get the name of a 'tiledb_profile' object}
\usage{
tiledb_profile_name(profile)
}
\value{
The name of the 'tiledb_profile' object
}
\description{
Get the name of a 'tiledb_profile' object
}
