% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_remove}
\alias{tiledb_profile_remove}
\title{Remove a saved 'tiledb_profile'}
\usage{
tiledb_profile_remove(name = NULL, dir = NULL)
}
\arguments{
\item{name}{(optional) Name of the profile to remove}

\item{dir}{(optional) Directory where the profile to remove is saved}
}
\description{
Remove a saved 'tiledb_profile'
}
