% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_save}
\alias{tiledb_profile_save}
\title{Save the 'tiledb_profile' object}
\usage{
tiledb_profile_save(profile)
}
\arguments{
\item{profile}{The 'tiledb_profile' object to save}
}
\description{
This will save the 'tiledb_profile' with the name and directory set at creation.
}
