% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile.R
\name{tiledb_profile_set_param}
\alias{tiledb_profile_set_param}
\title{Set a parameter on the 'tiledb_profile' object}
\usage{
tiledb_profile_set_param(profile, param, value)
}
\arguments{
\item{profile}{A TileDB profile object}

\item{param}{The key for the new parameter}

\item{value}{The value for the new parameter}
}
\description{
Set a parameter on the 'tiledb_profile' object
}
