% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_put_metadata}
\alias{tiledb_put_metadata}
\title{Store an object in TileDB Array Metadata under given key}
\usage{
tiledb_put_metadata(arr, key, val)
}
\arguments{
\item{arr}{A TileDB Array object, or a character URI describing one}

\item{key}{A character value describing a metadata key}

\item{val}{An object to be stored}
}
\value{
A boolean value indicating success
}
\description{
Store an object in TileDB Array Metadata under given key
}
