% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\docType{class}
\name{tiledb_query-class}
\alias{tiledb_query-class}
\title{An S4 class for a TileDB Query object}
\description{
An S4 class for a TileDB Query object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

