% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_add_range}
\alias{tiledb_query_add_range}
\title{Set a range for a given query}
\usage{
tiledb_query_add_range(query, schema, attr, lowval, highval, stride = NULL)
}
\arguments{
\item{query}{A TileDB Query object}

\item{schema}{A TileDB Schema object}

\item{attr}{An character variable with a dimension or attribute name}

\item{lowval}{The lower value of the range to be set for the query on this attribute}

\item{highval}{The highre value of the range to be set for the query on this attribute}

\item{stride}{An optional stride value for the range to be set for the query on this attribute}
}
\value{
The query object, invisibly
}
\description{
Set a range for a given query
}
