% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_apply_aggregate}
\alias{tiledb_query_apply_aggregate}
\title{Run an aggregate oprtation on the given query attribute}
\usage{
tiledb_query_apply_aggregate(
  query,
  attrname,
  operation = c("Count", "NullCount", "Min", "Max", "Mean", "Sum"),
  nullable = TRUE
)
}
\arguments{
\item{query}{A TileDB Query object}

\item{attrname}{The name of an attribute}

\item{operation}{The name of aggregation operation}

\item{nullable}{A boolean toggle whether the attribute is nullable}
}
\value{
The value of the aggregation
}
\description{
Run an aggregate oprtation on the given query attribute
}
