% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\docType{class}
\name{tiledb_query_condition-class}
\alias{tiledb_query_condition-class}
\title{An S4 class for a TileDB QueryCondition object}
\description{
An S4 class for a TileDB QueryCondition object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}

\item{\code{init}}{A logical variable tracking if the query condition object has been
initialized}
}}

