% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition}
\alias{tiledb_query_condition}
\title{Creates a 'tiledb_query_condition' object}
\usage{
tiledb_query_condition(ctx = tiledb_get_context())
}
\arguments{
\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A 'tiledb_query_condition' object
}
\description{
Creates a 'tiledb_query_condition' object
}
