% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition_combine}
\alias{tiledb_query_condition_combine}
\title{Combine two 'tiledb_query_condition' objects}
\usage{
tiledb_query_condition_combine(lhs, rhs, op)
}
\arguments{
\item{lhs}{A 'tiledb_query_condition' object on the left-hand side of the relation}

\item{rhs}{A 'tiledb_query_condition' object on the left-hand side of the relation}

\item{op}{A character value with then relation, this must be one of 'AND' or 'OR'.}
}
\value{
The combined 'tiledb_query_condition' object
}
\description{
Combines two query condition objects using a relational operator. Support for operator
'AND' is generally available. The 'OR' operator is available in TileDB 2.10 and later.
}
