% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition_create}
\alias{tiledb_query_condition_create}
\title{create a query condition for vector 'IN' and 'NOT_IN' operations}
\usage{
tiledb_query_condition_create(
  name,
  values,
  op = "IN",
  ctx = tiledb_get_context()
)
}
\arguments{
\item{name}{A character value with the scheme attribute name}

\item{values}{A vector wiith the given values, supported types are integer, double,
integer64 and charactor}

\item{op}{(optional) A character value with the chosen set operation, this must be one of
\sQuote{IN} or \sQuote{NOT_IN}; default to \sQuote{IN}}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
A query condition object is returned
}
\description{
Uses \sQuote{IN} and \sQuote{NOT_IN} operators on given attribute
}
