% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition_negate}
\alias{tiledb_query_condition_negate}
\title{Negate a 'tiledb_query_condition' object}
\usage{
tiledb_query_condition_negate(qc)
}
\arguments{
\item{qc}{A 'tiledb_query_condition' object to negate}
}
\value{
The negated 'tiledb_query_condition' object
}
\description{
Takes a query condition object and negates it, similar to the '!' operator.
}
