% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_create_buffer_ptr}
\alias{tiledb_query_create_buffer_ptr}
\title{Allocate and populate a Query buffer for a given object of a given data type.}
\usage{
tiledb_query_create_buffer_ptr(query, datatype, object)
}
\arguments{
\item{query}{A TileDB Query object}

\item{datatype}{A character value containing the data type}

\item{object}{A vector object of the given type}
}
\value{
An external pointer to the allocated buffer object
}
\description{
This function allocates a query buffer for the given data object of the given
type and assigns the object content to the buffer.
}
