% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_create_buffer_ptr_char}
\alias{tiledb_query_create_buffer_ptr_char}
\title{Allocate and populate a Query buffer for writing the given char vector}
\usage{
tiledb_query_create_buffer_ptr_char(query, varvec)
}
\arguments{
\item{query}{A TileDB Query object}

\item{varvec}{A vector of strings}
}
\value{
An external pointer to the allocated buffer object
}
\description{
Allocate and populate a Query buffer for writing the given char vector
}
