% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_ctx}
\alias{tiledb_query_ctx}
\title{Return query context object}
\usage{
tiledb_query_ctx(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
A TileDB Context object retrieved from the query
}
\description{
Return query context object
}
