% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowio.R
\name{tiledb_query_export_buffer}
\alias{tiledb_query_export_buffer}
\title{Export Query Buffer to Pair of Arrow IO Pointers}
\usage{
tiledb_query_export_buffer(query, name)
}
\arguments{
\item{name}{A character variable identifying the buffer}

\item{queryxp}{An external pointer object to TileDB Query object}
}
\value{
A two-element numeric vector where the two elements are
pointer to the Arrow array and schema
}
\description{
This function exports the name buffer from \sQuote{READ} query
to two Arrow C pointers.
}
