% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_finalize}
\alias{tiledb_query_finalize}
\title{Finalize TileDB Query}
\usage{
tiledb_query_finalize(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
A character value, either 'READ' or 'WRITE'
}
\description{
Finalize TileDB Query
}
