% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_buffer_char}
\alias{tiledb_query_get_buffer_char}
\title{Retrieve content from a Query character buffer}
\usage{
tiledb_query_get_buffer_char(bufptr, sizeoffsets = 0, sizestring = 0)
}
\arguments{
\item{bufptr}{An external pointer with a query buffer}

\item{sizeoffsets}{An optional argument for the length of the internal offsets vector}

\item{sizestring}{An optional argument for the length of the internal string}
}
\value{
An R object as resulting from the query
}
\description{
This function uses a query buffer for a character attribute
or dimension and returns its content.
}
