% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_buffer_ptr}
\alias{tiledb_query_get_buffer_ptr}
\title{Retrieve content from a Query buffer}
\usage{
tiledb_query_get_buffer_ptr(bufptr)
}
\arguments{
\item{bufptr}{An external pointer with a query buffer}
}
\value{
An R object as resulting from the query
}
\description{
This function uses a query buffer and returns its content.
}
