% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_est_result_size}
\alias{tiledb_query_get_est_result_size}
\title{Retrieve the estimated result size for a query and attribute}
\usage{
tiledb_query_get_est_result_size(query, name)
}
\arguments{
\item{query}{A TileDB Query object}

\item{name}{A variable with an attribute name}
}
\value{
An estimate of the query result size
}
\description{
When reading from sparse arrays, one cannot know beforehand how big the
result will be (unless one actually executes the query). This function
offers a way to get the estimated result size for the given attribute.
As TileDB does not actually execute the query, getting the estimated
result is very fast.
}
