% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_fragment_num}
\alias{tiledb_query_get_fragment_num}
\title{Retrieve the Number of Fragments for Query}
\usage{
tiledb_query_get_fragment_num(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
An integer with the number of fragments for the given query
}
\description{
This function is only applicable to \sQuote{WRITE} queries.
}
