% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_fragment_uri}
\alias{tiledb_query_get_fragment_uri}
\title{Retrieve the URI for a given Query Fragment}
\usage{
tiledb_query_get_fragment_uri(query, idx)
}
\arguments{
\item{query}{A TileDB Query object}

\item{idx}{An integer (or numeric) index ranging from zero to the number of fragments minus 1}
}
\value{
An character value with the fragment URI
}
\description{
This function is only applicable to \sQuote{WRITE} queries.
}
