% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_layout}
\alias{tiledb_query_get_layout}
\title{Get TileDB Query layout}
\usage{
tiledb_query_get_layout(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
The TileDB Query layout as a string
}
\description{
Get TileDB Query layout
}
