% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_range_num}
\alias{tiledb_query_get_range_num}
\title{Retrieve the number of ranges for a query dimension}
\usage{
tiledb_query_get_range_num(query, idx)
}
\arguments{
\item{query}{A TileDB Query object}

\item{idx}{An integer or numeric index selecting the dimension}
}
\value{
An integer with the number of query range for the
given dimensions
}
\description{
Retrieve the number of ranges for a query dimension
}
