% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_range_var}
\alias{tiledb_query_get_range_var}
\title{Retrieve the query range for a variable-sized query dimension and range index}
\usage{
tiledb_query_get_range_var(query, dimidx, rngidx)
}
\arguments{
\item{query}{A TileDB Query object}

\item{dimidx}{An integer index selecting the variable-sized dimension}

\item{rngidx}{An integer index selection the given range for the dimension}
}
\value{
An string vector with elements start and end for the query
range for the given dimension and range index
}
\description{
Retrieve the query range for a variable-sized query dimension and range index
}
