% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_result_buffer_elements_vec}
\alias{tiledb_query_result_buffer_elements_vec}
\title{Get TileDB Query result buffer element size pair as vector}
\usage{
tiledb_query_result_buffer_elements_vec(query, attr)
}
\arguments{
\item{query}{A TileDB Query object}

\item{attr}{A character value containing the attribute}
}
\value{
A vector with the number of elements in the offsets buffer (and zero
for fixed-size attribute or dimensions) and the number elements in the results
buffer #' for the given attribute
}
\description{
The underlying library functions returns a pair of values as a vector
of length two. The first number is the number of element offsets for variable
size attributes (and always zero for fixed-sized attributes and coordinates).
The second is the number of elements in the data buffer. For variable-sized
attributes the first number is the number of cells read (and hence the number
of offsets), the second number is the number of elements in the data buffer.
}
\seealso{
tiledb_query_result_buffer_elements
}
