% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_set_buffer_ptr}
\alias{tiledb_query_set_buffer_ptr}
\title{Assigns to a Query buffer for a given attribute}
\usage{
tiledb_query_set_buffer_ptr(query, attr, bufptr)
}
\arguments{
\item{query}{A TileDB Query object}

\item{attr}{A character value containing the attribute}

\item{bufptr}{An external pointer with a query buffer}
}
\value{
The modified query object, invisibly
}
\description{
This function assigns a given query buffer to a query.
}
