% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_set_buffer_ptr_var}
\alias{tiledb_query_set_buffer_ptr_var}
\title{Assign a buffer to a Query attribute}
\usage{
tiledb_query_set_buffer_ptr_var(query, attr, bufptr)
}
\arguments{
\item{query}{A TileDB Query object}

\item{attr}{A character value containing the attribute}

\item{bufptr}{An external pointer with a query buffer}
}
\value{
The modified query object, invisibly
}
\description{
Assign a buffer to a Query attribute
}
