% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_set_condition}
\alias{tiledb_query_set_condition}
\title{Set a query combination object for a query}
\usage{
tiledb_query_set_condition(query, qc)
}
\arguments{
\item{query}{A TileDB Query object}

\item{qc}{A TileDB Query Combination object}
}
\value{
The modified query object, invisibly
}
\description{
Set a query combination object for a query
}
