% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_set_layout}
\alias{tiledb_query_set_layout}
\title{Set TileDB Query layout}
\usage{
tiledb_query_set_layout(
  query,
  layout = c("COL_MAJOR", "ROW_MAJOR", "GLOBAL_ORDER", "UNORDERED")
)
}
\arguments{
\item{query}{A TileDB Query object}

\item{layout}{A character variable with the layout; must be one of
"COL_MAJOR", "ROW_MAJOR", "GLOBAL_ORDER", "UNORDERED")}
}
\value{
The modified query object, invisibly
}
\description{
Set TileDB Query layout
}
