% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_set_subarray}
\alias{tiledb_query_set_subarray}
\title{Set subarray for TileDB Query object}
\usage{
tiledb_query_set_subarray(query, subarray, type)
}
\arguments{
\item{query}{A TileDB Query object}

\item{subarray}{A subarry vector object}

\item{type}{An optional type as a character, if missing type is
inferred from the vector.}
}
\value{
The modified query object, invisibly
}
\description{
Set subarray for TileDB Query object
}
