% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_stats}
\alias{tiledb_query_stats}
\title{Return query statistics as a JSON string}
\usage{
tiledb_query_stats(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
A JSON-formatted string with context statistics
}
\description{
Return query statistics as a JSON string
}
