% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_status}
\alias{tiledb_query_status}
\title{Get TileDB Query status}
\usage{
tiledb_query_status(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
A character value describing the query status
}
\description{
Get TileDB Query status
}
