% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_submit_async}
\alias{tiledb_query_submit_async}
\title{Submit TileDB Query asynchronously without a callback returning immediately}
\usage{
tiledb_query_submit_async(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
The modified query object, invisibly
}
\description{
Note that the query object may need to be finalized
via \code{tiledb_query_finalize}.
}
