% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_type}
\alias{tiledb_query_type}
\title{Return TileDB Query type}
\usage{
tiledb_query_type(query)
}
\arguments{
\item{query}{A TileDB Query object}
}
\value{
A character value, either 'READ' or 'WRITE'
}
\description{
Return TileDB Query type
}
