% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_schema_get_dim_attr_status}
\alias{tiledb_schema_get_dim_attr_status}
\title{Get Dimension or Attribute Status}
\usage{
tiledb_schema_get_dim_attr_status(sch)
}
\arguments{
\item{sch}{A TileDB Schema object}
}
\value{
An integer vector where each element corresponds to a schema entry,
and a value of one signals dimension and a value of two an attribute.
}
\description{
Note that this function is an unexported internal function.
}
