% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_schema_get_names}
\alias{tiledb_schema_get_names}
\title{Get all Dimension and Attribute Names}
\usage{
tiledb_schema_get_names(sch)
}
\arguments{
\item{sch}{A TileDB Schema object}
}
\value{
A character vector of dimension and attribute names
}
\description{
Get all Dimension and Attribute Names
}
