% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_schema_get_types}
\alias{tiledb_schema_get_types}
\title{Get all Dimension and Attribute Types}
\usage{
tiledb_schema_get_types(sch)
}
\arguments{
\item{sch}{A TileDB Schema object}
}
\value{
A character vector of dimension and attribute data types
}
\description{
Get all Dimension and Attribute Types
}
