% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_set_vfs}
\alias{tiledb_set_vfs}
\title{Store a TileDB VFS object in the package environment}
\usage{
tiledb_set_vfs(vfs)
}
\arguments{
\item{vfs}{A TileDB VFS object}
}
\value{
NULL, invisibly. The function is invoked for the side-effect of
storing the VFS object.
}
\description{
Store a TileDB VFS object in the package environment
}
