% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{tiledb_sparse}
\alias{tiledb_sparse}
\title{Constructs a tiledb_sparse object backed by a persisted tiledb array uri}
\usage{
tiledb_sparse(ctx, uri, query_type = c("READ", "WRITE"),
  as.data.frame = FALSE)
}
\arguments{
\item{ctx}{tiledb_ctx}

\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}
}
\value{
tiledb_sparse array object
}
\description{
tiledb_sparse returns a list of coordinates and attributes vectors for reads
}
