% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{tiledb_stats_enable}
\alias{tiledb_stats_enable}
\title{Enable internal TileDB statistics counters}
\usage{
tiledb_stats_enable()
}
\description{
This function starts the collection of internal statistics.
}
