% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{tiledb_stats_raw_dump}
\alias{tiledb_stats_raw_dump}
\title{Dumps internal TileDB statistics as JSON to a string}
\usage{
tiledb_stats_raw_dump()
}
\description{
This function requires TileDB Embedded 2.0.3 or later.
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
if (tiledb_version(TRUE) >= "2.0.3") {
  txt <- tiledb_stats_raw_dump()
  cat(txt, "\n")
}
}
