% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{tiledb_stats_raw_get}
\alias{tiledb_stats_raw_get}
\title{Gets internal TileDB statistics as JSON string}
\usage{
tiledb_stats_raw_get()
}
\description{
This function is a convenience wrapper for \code{tiledb_stats_raw_dump}
and returns the result as a JSON string
}
