% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subarray.R
\docType{class}
\name{tiledb_subarray-class}
\alias{tiledb_subarray-class}
\title{An S4 class for a TileDB Subarray}
\description{
An S4 class for a TileDB Subarray
}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the underlying implementation}
}}

