% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subarray.R
\name{tiledb_subarray_to_query}
\alias{tiledb_subarray_to_query}
\title{Apply a Subarray to a Query}
\usage{
tiledb_subarray_to_query(query, subarray)
}
\arguments{
\item{query}{A TileDB Query Object}

\item{subarray}{A TileDB Subarray Object}
}
\value{
tiledb_query object
}
\description{
Apply a Subarray to a Query
}
