% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.R
\name{tiledb_version}
\alias{tiledb_version}
\title{The version of the libtiledb library}
\usage{
tiledb_version(compact = FALSE)
}
\arguments{
\item{compact}{Logical value indicating whether a compact
\code{package_version} object should be returned}
}
\value{
An named int vector c(major, minor, patch), or if select,
a \code{package_version} object
}
\description{
The version of the libtiledb library
}
\examples{
\dontshow{
ctx <- tiledb_ctx(limitTileDBCores())
}
tiledb_version()
tiledb_version(compact = TRUE)
}
