% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\docType{class}
\name{tiledb_vfs-class}
\alias{tiledb_vfs-class}
\title{An S4 class for a TileDB VFS object}
\description{
An S4 class for a TileDB VFS object
}
\section{Slots}{

\describe{
\item{\code{ptr}}{An external pointer to the underlying implementation}
}}

