% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs}
\alias{tiledb_vfs}
\title{Creates a \code{tiledb_vfs} object}
\usage{
tiledb_vfs(config = NULL, ctx = tiledb_get_context())
}
\arguments{
\item{config}{(optional) character vector of config parameter names, values}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
\code{tiledb_vfs} object
}
\description{
Creates a \code{tiledb_vfs} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
# default configuration
vfs <- tiledb_vfs()

}
