% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_close}
\alias{tiledb_vfs_close}
\title{Close a TileDB VFS Filehandle}
\usage{
tiledb_vfs_close(fh, ctx = tiledb_get_context())
}
\arguments{
\item{fh}{A TileDB VFS Filehandle external pointer as returned from \code{tiledb_vfs_open}}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
The result of the close operation is returned.
}
\description{
Close a TileDB VFS Filehandle
}
