% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_copy_dir}
\alias{tiledb_vfs_copy_dir}
\title{Copy a directory recursively to VFS}
\usage{
tiledb_vfs_copy_dir(dir, uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{dir}{Character variable with a local directory path}

\item{uri}{Character variable with a URI describing a directory path}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The uri value of the copied directory
}
\description{
Copy a directory recursively to VFS
}
