% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_copy_file}
\alias{tiledb_vfs_copy_file}
\title{Copy a file to VFS}
\usage{
tiledb_vfs_copy_file(file, uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{file}{Character variable with a local file path}

\item{uri}{Character variable with a URI describing a file path}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The uri value of the removed file
}
\description{
Copy a file to VFS
}
