% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_create_bucket}
\alias{tiledb_vfs_create_bucket}
\title{Create a VFS Bucket}
\usage{
tiledb_vfs_create_bucket(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a cloud bucket}
}
\value{
The uri value
}
\description{
Create a VFS Bucket
}
