% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_empty_bucket}
\alias{tiledb_vfs_empty_bucket}
\title{Empty a VFS Bucket}
\usage{
tiledb_vfs_empty_bucket(uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{uri}{Character variable with a URI describing a cloud bucket}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The URI value that was emptied
}
\description{
Empty a VFS Bucket
}
