% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_ls}
\alias{tiledb_vfs_ls}
\title{Return VFS Directory Listing}
\usage{
tiledb_vfs_ls(uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{uri}{Character variable with a URI describing a file path}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The content of the directory, non-recursive
}
\description{
Return VFS Directory Listing
}
