% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_ls_recursive}
\alias{tiledb_vfs_ls_recursive}
\title{Recursively list objects from given URI}
\usage{
tiledb_vfs_ls_recursive(
  uri,
  vfs = tiledb_get_vfs(),
  ctx = tiledb_get_context()
)
}
\arguments{
\item{uri}{Character variable with a URI describing a file path}

\item{vfs}{(optional) A TileDB VFS object; default is to use a cached value.}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
A data.frame object with two columns for the full path and the object
size in bytes
}
\description{
This functionality is currently limited to S3 URIs.
}
