% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_move_dir}
\alias{tiledb_vfs_move_dir}
\title{Move (or rename) a VFS Directory}
\usage{
tiledb_vfs_move_dir(vfs, olduri, newuri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{olduri}{Character variable with an existing URI describing a directory path}

\item{newuri}{Character variable with a new desired URI directory path}
}
\value{
The newuri value of the moved directory
}
\description{
Move (or rename) a VFS Directory
}
