% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_move_file}
\alias{tiledb_vfs_move_file}
\title{Move (or rename) a VFS File}
\usage{
tiledb_vfs_move_file(vfs, olduri, newuri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{olduri}{Character variable with an existing URI describing a file path}

\item{newuri}{Character variable with a new desired URI file path}
}
\value{
The newuri value of the moved file
}
\description{
Move (or rename) a VFS File
}
