% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_open}
\alias{tiledb_vfs_open}
\title{Open a TileDB VFS Filehandle for reading or writing}
\usage{
tiledb_vfs_open(
  binfile,
  mode = c("READ", "WRITE", "APPEND"),
  vfs = tiledb_get_vfs(),
  ctx = tiledb_get_context()
)
}
\arguments{
\item{binfile}{A character variable describing the (binary) file to be opened}

\item{mode}{A character variable with value \sQuote{READ}, \sQuote{WRITE}
or \sQuote{APPEND}}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
A TileDB VFS Filehandle object (as an external pointer)
}
\description{
Open a TileDB VFS Filehandle for reading or writing
}
