% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_remove_dir}
\alias{tiledb_vfs_remove_dir}
\title{Remove a VFS Directory}
\usage{
tiledb_vfs_remove_dir(uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{uri}{Character variable with a URI describing a diretory path}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The uri value of the removed directory
}
\description{
Remove a VFS Directory
}
