% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_remove_file}
\alias{tiledb_vfs_remove_file}
\title{Remove a VFS File}
\usage{
tiledb_vfs_remove_file(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a file path}
}
\value{
The uri value of the removed file
}
\description{
Remove a VFS File
}
