% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_sync}
\alias{tiledb_vfs_sync}
\title{Sync a TileDB VFS Filehandle}
\usage{
tiledb_vfs_sync(fh, ctx = tiledb_get_context())
}
\arguments{
\item{fh}{A TileDB VFS Filehandle external pointer as returned from \code{tiledb_vfs_open}}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
The result of the sync operation is returned.
}
\description{
Sync a TileDB VFS Filehandle
}
