% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_touch}
\alias{tiledb_vfs_touch}
\title{Touch a VFS URI Resource}
\usage{
tiledb_vfs_touch(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a bucket, file or directory}
}
\value{
The uri value
}
\description{
Touch a VFS URI Resource
}
