% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_unserialize}
\alias{tiledb_vfs_unserialize}
\title{Unserialize an R Object from a VFS-accessible URI}
\usage{
tiledb_vfs_unserialize(uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{uri}{Character variable with a URI describing a file path to an RDS file}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The unserialized object
}
\description{
Unserialize an R Object from a VFS-accessible URI
}
