% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_write}
\alias{tiledb_vfs_write}
\title{Write to a TileDB VFS Filehandle}
\usage{
tiledb_vfs_write(fh, vec, ctx = tiledb_get_context())
}
\arguments{
\item{fh}{A TileDB VFS Filehandle external pointer as returned from \code{tiledb_vfs_open}}

\item{vec}{An integer vector of content to be written}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
The result of the write operation is returned.
}
\description{
This interface currently defaults to using an integer vector. This is suitable for R objects
as the raw vector result from serialization can be mapped easily to an integer vector. It is
also possible to \code{memcpy} to the contiguous memory of an integer vector should other
(non-R) data be transferred.
}
