% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledbArray.R
\name{tiledbArray}
\alias{tiledbArray}
\title{Constructs a tiledbArray object backed by a persisted tiledb array uri}
\usage{
tiledbArray(
  uri,
  query_type = c("READ", "WRITE"),
  is.sparse = NA,
  as.data.frame = FALSE,
  attrs = character(),
  extended = TRUE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}

\item{is.sparse}{optional logical switch, defaults to "NA" letting array determine it}

\item{as.data.frame}{optional logical switch, defaults to "FALSE"}

\item{attrs}{optional character vector to select attributes, default is
empty implying all are selected}

\item{extended}{optional logical switch selecting wide \sQuote{data.frame}
format, defaults to "TRUE"}

\item{ctx}{tiledb_ctx (optional)}
}
\value{
tiledb_sparse array object
}
\description{
tiledbArray returns a new object. This class is experimental.
}
