% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{varlen_list_eltype}
\alias{varlen_list_eltype}
\title{Get R type of elements in a list}
\usage{
varlen_list_eltype(x)
}
\arguments{
\item{x}{a list}
}
\value{
single character, e.g. "numeric"
}
\description{
TileDB varlen arrays are implemented as R lists of arrays with a uniform storage mode.
}
