% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarlenArray.R
\name{varlen_offsets}
\alias{varlen_offsets}
\title{Calculate offsets}
\usage{
varlen_offsets(x)
}
\arguments{
\item{x}{An array or list (potentially with dimensions) of arrays}
}
\description{
Calculate offsets
}
