% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VariableLength.R
\name{write_variable_length}
\alias{write_variable_length}
\title{Write a variable-length array}
\usage{
write_variable_length(uri, listobject, debug = FALSE)
}
\arguments{
\item{uri}{A character variable with a URI for a (dense) array}

\item{listobject}{A named list with variable-length arrays.}

\item{debug}{An optional logical switch for increased verbosity, defaults
to FALSE.}
}
\value{
The result of the underlying write call is returned.
}
\description{
Using a given (named) list with variable-length arrays, write them out
to the given URI. Currently defaults to dense arrays.
}
