% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{array_consolidate}
\alias{array_consolidate}
\title{Consolidate fragments of a TileDB Array}
\usage{
array_consolidate(uri, cfg = NULL, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{A character value with the URI of a TileDB Array}

\item{cfg}{An optional TileDB Configuration object}

\item{ctx}{An option TileDB Context object}
}
\value{
\code{NULL} is returned invisibly
}
\description{
This function invokes a consolidation operation. Parameters can be set via
an option configuration object.
}
