% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{as.data.frame.tiledb_config}
\alias{as.data.frame.tiledb_config}
\title{Convert a \code{tiledb_config} object to a R data.frame}
\usage{
\method{as.data.frame}{tiledb_config}(x, ...)
}
\arguments{
\item{x}{\code{tiledb_config} object}

\item{...}{Extra parameter for method signature, currently unused.}
}
\value{
a data.frame wth parameter, value columns
}
\description{
Convert a \code{tiledb_config} object to a R data.frame
}
\examples{
cfg <- tiledb_config()
as.data.frame(cfg)

}
