% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{as.vector.tiledb_config}
\alias{as.vector.tiledb_config}
\title{Convert a TileDB Config object to a R vector}
\usage{
\method{as.vector}{tiledb_config}(x, mode = "any")
}
\arguments{
\item{x}{A \code{tiledb_config} object}

\item{mode}{A character value \code{"any"}, currently unused}
}
\value{
A character vector of config parameter names, values
}
\description{
Convert a TileDB Config object to a R vector
}
\examples{
\dontshow{
ctx <- tiledb_ctx(limitTileDBCores())
}
cfg <- tiledb_config()
as.vector(cfg)

}
