% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{attrs,tiledb_array_schema,ANY-method}
\alias{attrs,tiledb_array_schema,ANY-method}
\title{Returns a list of all \code{tiledb_attr} objects associated with the \code{tiledb_array_schema}}
\usage{
\S4method{attrs}{tiledb_array_schema,ANY}(object, idx, ...)
}
\arguments{
\item{object}{tiledb_array_schema}

\item{idx}{index argument, currently unused.}

\item{...}{Extra parameter for method signature, currently unused.}
}
\value{
a list of tiledb_attr objects
}
\description{
Returns a list of all \code{tiledb_attr} objects associated with the \code{tiledb_array_schema}
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 10L), type = "INT32")))
sch <- tiledb_array_schema(dom, attrs = c(tiledb_attr("a1", type = "INT32"),
                                               tiledb_attr("a2", type = "FLOAT64")))
attrs(sch)

lapply(attrs(sch), datatype)

}
