% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{config,tiledb_ctx-method}
\alias{config,tiledb_ctx-method}
\title{Retrieve the \code{tiledb_config} object from the \code{tiledb_ctx}}
\usage{
\S4method{config}{tiledb_ctx}(object = tiledb_get_context())
}
\arguments{
\item{object}{tiledb_ctx object}
}
\value{
\code{tiledb_config} object associated with the \code{tiledb_ctx} instance
}
\description{
Retrieve the \code{tiledb_config} object from the \code{tiledb_ctx}
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
ctx <- tiledb_ctx(c("sm.tile_cache_size" = "10"))
cfg <- config(ctx)
cfg["sm.tile_cache_size"]

}
