% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{datatype,tiledb_dim-method}
\alias{datatype,tiledb_dim-method}
\title{Return the \code{tiledb_dim} datatype}
\usage{
\S4method{datatype}{tiledb_dim}(object)
}
\arguments{
\item{object}{tiledb_dim object}
}
\value{
tiledb datatype string
}
\description{
Return the \code{tiledb_dim} datatype
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
d1 <- tiledb_dim("d1", domain = c(5L, 10L), tile = 2L, type = "INT32")
datatype(d1)

}
