% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{dim.tiledb_array_schema}
\alias{dim.tiledb_array_schema}
\title{Retrieve the dimension (domain extent) of the domain}
\usage{
\method{dim}{tiledb_array_schema}(x)
}
\arguments{
\item{x}{tiledb_array_schema}
}
\value{
a dimension vector
}
\description{
Only valid for integral (integer) domains
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 10L), type = "INT32")))
sch <- tiledb_array_schema(dom, attrs = c(tiledb_attr("a1", type = "INT32"),
                                               tiledb_attr("a2", type = "FLOAT64")))
dim(sch)

}
