% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{dimensions,tiledb_array_schema-method}
\alias{dimensions,tiledb_array_schema-method}
\title{Returns a list of \code{tiledb_dim} objects associated with the \code{tiledb_array_schema}}
\usage{
\S4method{dimensions}{tiledb_array_schema}(object)
}
\arguments{
\item{object}{tiledb_array_schema}
}
\value{
a list of tiledb_dim objects
}
\description{
Returns a list of \code{tiledb_dim} objects associated with the \code{tiledb_array_schema}
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 100L), type = "INT32"),
                              tiledb_dim("d2", c(1L, 50L), type = "INT32")))
sch <- tiledb_array_schema(dom, attrs = c(tiledb_attr("a1", type = "INT32")))
dimensions(dom)

lapply(dimensions(dom), name)

}
