% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R, R/Attribute.R, R/Ctx.R,
%   R/DenseArray.R, R/Dim.R, R/Domain.R, R/FilterList.R, R/generics.R
\name{domain}
\alias{domain}
\alias{dimensions}
\alias{attrs}
\alias{cell_order}
\alias{tile_order}
\alias{filter_list}
\alias{is.sparse}
\alias{tiledb_ndim}
\alias{name}
\alias{datatype}
\alias{cell_val_num}
\alias{config}
\alias{schema}
\alias{tile}
\alias{is.integral}
\alias{set_max_chunk_size}
\alias{max_chunk_size}
\alias{nfilters}
\alias{generics}
\title{Generic Methods}
\usage{
domain(object, ...)

dimensions(object, ...)

attrs(object, idx, ...)

cell_order(object, ...)

tile_order(object, ...)

filter_list(object, ...)

is.sparse(object, ...)

tiledb_ndim(object, ...)

name(object)

datatype(object)

cell_val_num(object)

config(object, ...)

schema(object, ...)

tile(object)

is.integral(object)

set_max_chunk_size(object, value)

max_chunk_size(object)

nfilters(object)
}
\arguments{
\item{object}{A TileDB object}

\item{...}{Variable argument}

\item{idx}{An index argument}

\item{value}{A value to be assigned}
}
\description{
Definition of generic methods
}
