% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R, R/Attribute.R, R/Ctx.R,
%   R/DenseArray.R, R/Dim.R, R/Domain.R, R/FilterList.R, R/TileDBArray.R,
%   R/generics.R
\name{raw_dump}
\alias{raw_dump}
\alias{domain}
\alias{dimensions}
\alias{attrs}
\alias{cell_order}
\alias{tile_order}
\alias{filter_list}
\alias{filter_list<-}
\alias{is.sparse}
\alias{tiledb_ndim}
\alias{name}
\alias{datatype}
\alias{config}
\alias{schema}
\alias{tile}
\alias{is.integral}
\alias{nfilters}
\alias{tdb_filter}
\alias{tdb_select}
\alias{tdb_collect}
\alias{generics}
\title{Generic Methods}
\usage{
raw_dump(object, ...)

domain(object, ...)

dimensions(object, ...)

attrs(object, idx, ...)

cell_order(object, ...)

tile_order(object, ...)

filter_list(object, ...)

filter_list(x) <- value

is.sparse(object, ...)

tiledb_ndim(object, ...)

name(object)

datatype(object)

config(object, ...)

schema(object, ...)

tile(object)

is.integral(object)

nfilters(object)

tdb_filter(x, ...)

tdb_select(x, ...)

tdb_collect(x, ...)
}
\arguments{
\item{object}{A TileDB object}

\item{...}{Variable argument}

\item{idx}{An index argument}

\item{x}{A TileDB Object}

\item{value}{A value to be assigned}
}
\description{
Definition of generic methods
}
