% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{is.sparse,tiledb_dense-method}
\alias{is.sparse,tiledb_dense-method}
\title{Returns true is if the array or array_schema is sparse}
\usage{
\S4method{is.sparse}{tiledb_dense}(object)
}
\arguments{
\item{object}{tiledb_dense}
}
\value{
FALSE
}
\description{
Returns true is if the array or array_schema is sparse
}
