% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\docType{methods}
\name{name,tiledb_attr-method}
\alias{name,tiledb_attr-method}
\title{Return the \code{tiledb_attr} name}
\usage{
\S4method{name}{tiledb_attr}(object)
}
\arguments{
\item{`tiledb_attr`}{object}
}
\value{
string name, empty string if the attribute is anonymous
}
\description{
Return the \code{tiledb_attr} name
}
\examples{
ctx <- tiledb_ctx()
a1 <- tiledb_attr(ctx, "a1")
name(a1)

a2 <- tiledb_attr(ctx)
name(a2)

}
