% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{name,tiledb_attr-method}
\alias{name,tiledb_attr-method}
\title{Return the \code{tiledb_attr} name}
\usage{
\S4method{name}{tiledb_attr}(object)
}
\arguments{
\item{object}{\code{tiledb_attr} object}
}
\value{
string name, empty string if the attribute is anonymous
}
\description{
Return the \code{tiledb_attr} name
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
a1 <- tiledb_attr("a1", type = "INT32")
name(a1)

a2 <- tiledb_attr(type = "INT32")
name(a2)

}
