% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{show,tiledb_domain-method}
\alias{show,tiledb_domain-method}
\title{Prints an domain object}
\usage{
\S4method{show}{tiledb_domain}(object)
}
\arguments{
\item{object}{An domain object}
}
\description{
Prints an domain object
}
