% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_array.R
\name{[,tiledb_array,ANY-method}
\alias{[,tiledb_array,ANY-method}
\title{Gets a tiledb_array value}
\usage{
\S4method{[}{tiledb_array,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{tiledb_array object}

\item{i}{row index expression}

\item{j}{column index expression}

\item{...}{Extra parameter for method signature}

\item{drop}{Optional logical switch to drop dimensions, default FALSE, currently unused.}
}
\value{
An element from the sparse array
}
\description{
This function is incomplete.
}
