% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterList.R
\name{[,tiledb_filter_list-method}
\alias{[,tiledb_filter_list-method}
\title{Returns the filter at given index}
\usage{
\S4method{[}{tiledb_filter_list}(x, i, j, ..., drop = TRUE)
}
\value{
object tiledb_filter
}
\description{
Returns the filter at given index
}
\examples{
flt <- tiledb_filter("ZSTD")
tiledb_filter_set_option(flt, "COMPRESSION_LEVEL", 5)
filter_list <- tiledb_filter_list(c(flt))
filter_list[0]

}
