% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{tdb_filter,tiledb_array-method}
\alias{tdb_filter,tiledb_array-method}
\title{Filter from array for query via logical conditions}
\usage{
\S4method{tdb_filter}{tiledb_array}(x, ..., strict = TRUE)
}
\arguments{
\item{x}{A tiledb_array object as first argument, permitting piping}

\item{...}{One or more expressions that are parsed as query_condition objects}

\item{strict}{A boolean toogle to, if set, errors if a non-existing attribute is selected
or filtered on, defaults to 'TRUE'; if 'FALSE' a warning is shown by execution proceeds.}
}
\value{
The tiledb_array object, permitting piping
}
\description{
Filter from array for query via logical conditions
}
