% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{tile,tiledb_dim-method}
\alias{tile,tiledb_dim-method}
\title{Return the \code{tiledb_dim} tile extent}
\usage{
\S4method{tile}{tiledb_dim}(object)
}
\arguments{
\item{object}{\code{tiledb_dim} object}
}
\value{
a scalar tile extent
}
\description{
Return the \code{tiledb_dim} tile extent
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
d1 <- tiledb_dim("d1", domain = c(5L, 10L), tile = 2L)
tile(d1)

}
