% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb-package.R
\docType{package}
\name{tiledb-package}
\alias{tiledb-package}
\title{tiledb - Interface to the TileDB Storage Manager API}
\description{
The efficient multi-dimensional array management system
'TileDB' introduces a novel on-disk format that can effectively store
reads. It features excellent compression, an efficient parallel I/O
system which also scales well, and bindings to multiple languages.
}
