% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_array.R
\docType{class}
\name{tiledb_array-class}
\alias{tiledb_array-class}
\title{An S4 class for a TileDB Array}
\description{
This class is experimental.
}
\section{Slots}{

\describe{
\item{\code{ctx}}{A TileDB context object}

\item{\code{uri}}{A character despription}

\item{\code{is.sparse}}{A logical value}

\item{\code{as.data.frame}}{A logical value}

\item{\code{attrs}}{A character vector}

\item{\code{extended}}{A logical value}

\item{\code{ptr}}{External pointer to the underlying implementation}
}}

