% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\docType{class}
\name{tiledb_array-class}
\alias{tiledb_array-class}
\title{An S4 class for a TileDB Array}
\description{
This class aims to eventually replace \code{\link{tiledb_dense}}
and \code{\link{tiledb_sparse}} provided equivalent functionality
based on refactored implementation utilising newer TileDB features.
}
\section{Slots}{

\describe{
\item{\code{ctx}}{A TileDB context object}

\item{\code{uri}}{A character despription}

\item{\code{is.sparse}}{A logical value}

\item{\code{as.data.frame}}{A logical value}

\item{\code{attrs}}{A character vector}

\item{\code{extended}}{A logical value}

\item{\code{selected_ranges}}{An optional list with matrices where each matrix i
describes the (min,max) pair of ranges for dimension i}

\item{\code{query_layout}}{An optional character value}

\item{\code{datetimes_as_int64}}{A logical value}

\item{\code{encryption_key}}{A character value}

\item{\code{timestamp}}{A POSIXct datetime variable}

\item{\code{as.matrix}}{A logical value}

\item{\code{as.array}}{A logical value}

\item{\code{query_condition}}{A Query Condition object}

\item{\code{ptr}}{External pointer to the underlying implementation}
}}

