% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArray.R
\name{tiledb_array}
\alias{tiledb_array}
\title{Constructs a tiledb_array object backed by a persisted tiledb array uri}
\usage{
tiledb_array(
  uri,
  query_type = c("READ", "WRITE"),
  is.sparse = NA,
  as.data.frame = FALSE,
  attrs = character(),
  extended = TRUE,
  selected_ranges = list(),
  ctx = tiledb_get_context()
)
}
\arguments{
\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}

\item{is.sparse}{optional logical switch, defaults to "NA" letting array determine it}

\item{as.data.frame}{optional logical switch, defaults to "FALSE"}

\item{attrs}{optional character vector to select attributes, default is
empty implying all are selected}

\item{extended}{optional logical switch selecting wide \sQuote{data.frame}
format, defaults to "TRUE"}

\item{selected_ranges}{An optional list with matrices where each matrix i
describes the (min,max) pair of ranges for dimension i}

\item{ctx}{tiledb_ctx (optional)}
}
\value{
tiledb_array object
}
\description{
tiledb_array returns a new object. This class is experimental.
}
