% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Array.R
\name{tiledb_array_open}
\alias{tiledb_array_open}
\title{Open a TileDB Array}
\usage{
tiledb_array_open(arr, type = c("READ", "WRITE"))
}
\arguments{
\item{arr}{A TileDB Array object as for example returned by \code{tiledb_dense()}}

\item{type}{A character value that must be either \sQuote{READ} or \sQuote{WRITE}}
}
\value{
The TileDB Array object but opened for reading or writing
}
\description{
Open a TileDB Array
}
