% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_array_schema_get_current_domain}
\alias{tiledb_array_schema_get_current_domain}
\title{Get the Current Domain of an Array Schema}
\usage{
tiledb_array_schema_get_current_domain(schema, ctx = tiledb_get_context())
}
\arguments{
\item{schema}{An Array Schema}

\item{ctx}{Optional tiledb_ctx object}
}
\value{
A 'CurrendDomain' object
}
\description{
Note that 'CurrendDomain' object may be empty.
}
