% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attr}
\alias{tiledb_attr}
\title{Contructs a \code{tiledb_attr} object}
\usage{
tiledb_attr(ctx = tiledb:::ctx, name = "", type = "FLOAT64",
  filter_list = tiledb_filter_list(ctx), ncells = 1)
}
\arguments{
\item{name}{(default "") The dimension name / label string.}

\item{type}{(default "FLOAT64") The tiledb_attr TileDB datatype string}

\item{filter_list}{(default filter_list("NONE")) The tiledb_attr filter_list}

\item{ncells}{(default 1) The number of cells}
}
\value{
\code{tiledb_dim} object
}
\description{
Contructs a \code{tiledb_attr} object
}
\examples{
ctx <- tiledb_ctx()
flt = filter_list("GZIP")
attr <- tiledb_attr(ctx, name = "a1", type = "INT32",
                    filter_list = flt)
attr

}
