% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute.R
\name{tiledb_attribute_get_enumeration}
\alias{tiledb_attribute_get_enumeration}
\alias{tiledb_attribute_get_enumeration_ptr}
\title{Get the TileDB Attribute Enumeration}
\usage{
tiledb_attribute_get_enumeration(attr, arr, ctx = tiledb_get_context())

tiledb_attribute_get_enumeration_ptr(attr, arrptr, ctx = tiledb_get_context())
}
\arguments{
\item{attr}{A TileDB Attribute object}

\item{arr}{A Tiledb Array object}

\item{ctx}{A Tiledb Context object (optional)}

\item{arrptr}{A Tiledb Array object pointer}
}
\value{
A character vector with the enumeration (of length zero if none)
}
\description{
Get the TileDB Attribute Enumeration
}
