% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config}
\alias{tiledb_config}
\title{Creates a \code{tiledb_config} object}
\usage{
tiledb_config(config = NA_character_)
}
\arguments{
\item{config}{(optional) character vector of config parameter names, values}
}
\value{
\code{tiledb_config} object
}
\description{
Creates a \code{tiledb_config} object
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
cfg <- tiledb_config()
cfg["sm.tile_cache_size"]

# set tile cache size to custom value
cfg <- tiledb_config(c("sm.tile_cache_size" = "100"))
cfg["sm.tile_cache_size"]

}
