% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Config.R
\name{tiledb_config_load}
\alias{tiledb_config_load}
\title{Load a saved \code{tiledb_config} file from disk}
\usage{
tiledb_config_load(path)
}
\arguments{
\item{path}{path to the config file}
}
\description{
Load a saved \code{tiledb_config} file from disk
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
tmp <- tempfile()
cfg <- tiledb_config(c("sm.tile_cache_size" = "10"))
pth <- tiledb_config_save(cfg, tmp)
cfg <- tiledb_config_load(pth)
cfg["sm.tile_cache_size"]

}
