% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DenseArray.R
\name{tiledb_dense}
\alias{tiledb_dense}
\title{Constructs a tiledb_dense object backed by a persisted tiledb array uri}
\usage{
tiledb_dense(
  uri,
  query_type = c("READ", "WRITE"),
  as.data.frame = FALSE,
  attrs = character(),
  extended = FALSE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}

\item{as.data.frame}{optional logical switch, defaults to "FALSE"}

\item{attrs}{optional character vector to select attributes, default is
empty implying all are selected}

\item{extended}{optional logical switch selecting wide \sQuote{data.frame}
format, defaults to "FALSE"}

\item{ctx}{tiledb_ctx (optional)}
}
\value{
tiledb_dense array object
}
\description{
Constructs a tiledb_dense object backed by a persisted tiledb array uri
}
