% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_domain}
\alias{tiledb_domain}
\title{Constructs a \code{tiledb_domain} object}
\usage{
tiledb_domain(dims, ctx = tiledb_get_context())
}
\arguments{
\item{dims}{list() of tiledb_dim objects}

\item{ctx}{tiledb_ctx (optional)}
}
\value{
tiledb_domain
}
\description{
All \code{tiledb_dim} must be of the same TileDB type.
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(1L, 100L), type = "INT32"),
                              tiledb_dim("d2", c(1L, 50L), type = "INT32")))
}
