% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{tiledb_filter}
\alias{tiledb_filter}
\title{Constructs a \code{tiledb_filter} object}
\usage{
tiledb_filter(name = "NONE", ctx = tiledb_get_context())
}
\arguments{
\item{name}{(default "NONE") TileDB filter name string}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
tiledb_filter object
}
\description{
Available filters:
\itemize{
\item "NONE"
\item "GZIP"
\item "ZSTD"
\item "LZ4"
\item "RLE"
\item "BZIP2"
\item "DOUBLE_DELTA"
\item "BIT_WIDTH_REDUCTION"
\item "BITSHUFFLE"
\item "BYTESHUFFLE"
\item "POSITIVE_DELTA"
\item "CHECKSUM_MD5"
\item "CHECKSUM_SHA256"
\item "DICTIONARY"
\item "SCALE_FLOAT"  (TileDB 2.11.0 or later)
}
}
\details{
Valid compression options vary depending on the filter used,
consult the TileDB docs for more information.
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
tiledb_filter("ZSTD")

}
