% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{tiledb_filter_set_option}
\alias{tiledb_filter_set_option}
\title{Set the filter's option}
\usage{
tiledb_filter_set_option(object, option, value)
}
\arguments{
\item{object}{tiledb_filter}

\item{option}{string}

\item{value}{int}
}
\description{
Set the filter's option
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
c <- tiledb_filter("ZSTD")
tiledb_filter_set_option(c,"COMPRESSION_LEVEL", 5)
tiledb_filter_get_option(c, "COMPRESSION_LEVEL")
}
