% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{tiledb_filter_type}
\alias{tiledb_filter_type}
\title{Returns the type of the filter used}
\usage{
tiledb_filter_type(object)
}
\arguments{
\item{object}{tiledb_filter}
}
\value{
TileDB filter type string
}
\description{
Returns the type of the filter used
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
c <- tiledb_filter("ZSTD")
tiledb_filter_type(c)

}
