% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_non_empty_domain_name}
\alias{tiledb_fragment_info_get_non_empty_domain_name}
\title{Return a fragment info non-empty domain from name}
\usage{
tiledb_fragment_info_get_non_empty_domain_name(object, fid, dim_name, typestr)
}
\arguments{
\item{object}{A TileDB fragment info object}

\item{fid}{A fragment object index}

\item{dim_name}{A character variable with the dimension name}

\item{typestr}{An optional character variable describing the data type which will
be accessed from the schema if missinh}
}
\value{
A TileDB Domain object
}
\description{
TODO: Rework with type information
}
