% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FragmentInfo.R
\name{tiledb_fragment_info_get_num}
\alias{tiledb_fragment_info_get_num}
\title{Return a fragment info number of fragments}
\usage{
tiledb_fragment_info_get_num(object)
}
\arguments{
\item{object}{A TileDB fragment info object}
}
\value{
A numeric variable with the number of fragments
}
\description{
Return a fragment info number of fragments
}
