% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metadata.R
\name{tiledb_get_all_metadata}
\alias{tiledb_get_all_metadata}
\title{Return all TileDB Array Metadata objects as a named list}
\usage{
tiledb_get_all_metadata(arr)
}
\arguments{
\item{arr}{A TileDB Array object}
}
\value{
A named list with all Metadata objects indexed by the given key
}
\description{
Return all TileDB Array Metadata objects as a named list
}
