% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R, R/Object.R
\name{tiledb_group_create}
\alias{tiledb_group_create}
\title{Create a TileDB Group}
\usage{
tiledb_group_create(uri, ctx = tiledb_get_context())

tiledb_group_create(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{path which to create group}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
The uri path, invisibly

uri of created group
}
\description{
Create a TileDB Group

Creates a TileDB group object at given uri path
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
\dontrun{
pth <- tempdir()
tiledb_group_create(pth)
tiledb_object_type(pth)
}
}
