% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group.R
\name{tiledb_group_create}
\alias{tiledb_group_create}
\title{Create a TileDB Group at the given path}
\usage{
tiledb_group_create(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{Character variable with the URI of the new group}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
The uri path, invisibly
}
\description{
Create a TileDB Group at the given path
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
\dontrun{
pth <- tempdir()
tiledb_group_create(pth)
tiledb_object_type(pth)
}
}
