% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ctx.R
\name{tiledb_is_supported_fs}
\alias{tiledb_is_supported_fs}
\title{Query if a TileDB backend is supported}
\usage{
tiledb_is_supported_fs(scheme, object = tiledb:::ctx)
}
\arguments{
\item{object}{\code{tiledb_ctx} object}

\item{schema}{URI string scheme ("file", "hdfs", "s3")}
}
\value{
TRUE if tiledb backend is supported, FALSE otherwise
}
\description{
The scheme corresponds to the URI scheme for TileDB resouces.
}
\details{
Ex:
\itemize{
\item \code{{file}:///path/to/file}
\item \code{{hdfs}:///path/to/file}
\item \code{{s3}://hostname:port/path/to/file}
}
}
\examples{
tiledb_is_supported_fs("file")
tiledb_is_supported_fs("s3")

}
