% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dim.R
\name{tiledb_ndim,tiledb_dim-method}
\alias{tiledb_ndim,tiledb_dim-method}
\title{Returns the number of dimensions for a tiledb domain object}
\usage{
\S4method{tiledb_ndim}{tiledb_dim}(object)
}
\arguments{
\item{object}{A \code{tiledb_dim} object}
}
\value{
An integer with the number of dimensions.
}
\description{
Returns the number of dimensions for a tiledb domain object
}
\examples{
\dontshow{
ctx <- tiledb_ctx(limitTileDBCores())
}
d1 <- tiledb_dim("d1", c(1L, 10L), 10L)
tiledb_ndim(d1)

}
