% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Domain.R
\name{tiledb_ndim,tiledb_domain-method}
\alias{tiledb_ndim,tiledb_domain-method}
\title{Returns the number of dimensions of the \code{tiledb_domain}}
\usage{
\S4method{tiledb_ndim}{tiledb_domain}(object)
}
\arguments{
\item{object}{tiledb_domain}
}
\value{
integer number of dimensions
}
\description{
Returns the number of dimensions of the \code{tiledb_domain}
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(0.5, 100.0), type = "FLOAT64")))
tiledb_ndim(dom)
dom <- tiledb_domain(dims = c(tiledb_dim("d1", c(0.5, 100.0), type = "FLOAT64"),
                                   tiledb_dim("d2", c(0.5, 100.0), type = "FLOAT64")))
tiledb_ndim(dom)

}
