% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Object.R
\name{tiledb_object_type}
\alias{tiledb_object_type}
\title{Return the TileDB object type string of a TileDB resource}
\usage{
tiledb_object_type(uri, ctx = tiledb_get_context())
}
\arguments{
\item{uri}{path to TileDB resource}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
TileDB object type string
}
\description{
Object types:
\itemize{
\item \code{"ARRAY"}, dense or sparse TileDB array
\item \code{"GROUP"}, TileDB group
\item `"INVALID"``, not a TileDB resource
}
}
