% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query}
\alias{tiledb_query}
\title{Creates a 'tiledb_query' object}
\usage{
tiledb_query(
  array,
  type = if (tiledb_version(TRUE) >= "2.12.0") c("READ", "WRITE", "DELETE",
    "MODIFY_EXCLUSIVE") else c("READ", "WRITE"),
  ctx = tiledb_get_context()
)
}
\arguments{
\item{array}{A TileDB Array object}

\item{type}{A character value that must be one of 'READ', 'WRITE', or
'DELETE' (for TileDB >= 2.12.0)}

\item{ctx}{(optional) A TileDB Ctx object}
}
\value{
'tiledb_query' object
}
\description{
Creates a 'tiledb_query' object
}
