% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_alloc_buffer_ptr_char_subarray}
\alias{tiledb_query_alloc_buffer_ptr_char_subarray}
\title{Allocate a Query buffer for reading a character attribute using a subarray}
\usage{
tiledb_query_alloc_buffer_ptr_char_subarray(
  array,
  attr,
  subarray = NULL,
  sizeoffsets = 0,
  sizedata = 0
)
}
\arguments{
\item{array}{A TileDB Array object}

\item{attr}{A character value containing the attribute}

\item{subarray}{A vector of length four describing the subarray required for dense arrays}

\item{sizeoffsets}{An optional value of the size of the offsets vector}

\item{sizedata}{An optional value of the size of the data string}
}
\value{
An external pointer to the allocated buffer object
}
\description{
Note that this uses an API part that may be deprecated in the future.
}
