% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition_combine}
\alias{tiledb_query_condition_combine}
\title{Combine two 'tiledb_query_condition' objects}
\usage{
tiledb_query_condition_combine(lhs, rhs, op)
}
\arguments{
\item{lhs}{A 'tiledb_query_condition' object on the left-hand side of the relation}

\item{rhs}{A 'tiledb_query_condition' object on the left-hand side of the relation}

\item{op}{A character value with then relation, this must be one of 'AND', 'OR' or 'NOT'.}
}
\value{
The combined 'tiledb_query_condition' object
}
\description{
Combines two query condition object using a relatiional operator. Note that at present
only 'AND' is supported.
}
