% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QueryCondition.R
\name{tiledb_query_condition_set_use_enumeration}
\alias{tiledb_query_condition_set_use_enumeration}
\title{Enable use of enumeration in query condition}
\usage{
tiledb_query_condition_set_use_enumeration(
  qc,
  use_enum,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{qc}{A 'tiledb_query_condition' object}

\item{use_enum}{A boolean to set (if TRUE) or unset (if FALSE) enumeration use}

\item{ctx}{(optional) A TileDB Ctx object; if not supplied the default
context object is retrieved}
}
\value{
Nothing is retuned, the function is invoked for the side effect
}
\description{
Set a boolean toggle to signal use of enumeration in query condtion (TileDB 2.17 or later)
}
