% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowIO.R
\name{tiledb_query_export_buffer}
\alias{tiledb_query_export_buffer}
\title{Export Query Buffer to Pair of Arrow IO Pointers}
\usage{
tiledb_query_export_buffer(query, name, ctx = tiledb_get_context())
}
\arguments{
\item{query}{A TileDB Query object}

\item{name}{A character variable identifying the buffer}

\item{ctx}{tiledb_ctx object (optional)}
}
\value{
A two-element numeric vector where the two elements are
pointers to the Arrow array and schema
}
\description{
This function exports the named buffer from a \sQuote{READ} query
to two Arrow C pointers.
}
