% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_get_fragment_timestamp_range}
\alias{tiledb_query_get_fragment_timestamp_range}
\title{Retrieve the timestamp range for a given Query Fragment}
\usage{
tiledb_query_get_fragment_timestamp_range(query, idx)
}
\arguments{
\item{query}{A TileDB Query object}

\item{idx}{An integer (or numeric) index ranging from zero to the number of fragments minus 1}
}
\value{
A two-element datetime vector with the start and end time of the fragment write.
}
\description{
This function is only applicable to \sQuote{WRITE} queries. The time resolution in
TileDB is millseconds since the epoch so an R \code{Datetime} vector is returned.
}
