% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArrowIO.R
\name{tiledb_query_import_buffer}
\alias{tiledb_query_import_buffer}
\title{Import to Query Buffer from Pair of Arrow IO Pointers}
\usage{
tiledb_query_import_buffer(query, name, arrowpointers)
}
\arguments{
\item{query}{A TileDB Query object}

\item{name}{A character variable identifying the buffer}

\item{arrowpointers}{A two-element numeric vector with two pointers
to an Arrow Array and Schema, respectively}
}
\value{
The update Query external pointer is returned
}
\description{
This function imports to the named buffer for a \sQuote{WRITE} query
from two Arrow C pointers.
}
