% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Query.R
\name{tiledb_query_result_buffer_elements}
\alias{tiledb_query_result_buffer_elements}
\title{Get TileDB Query result buffer elements}
\usage{
tiledb_query_result_buffer_elements(query, attr)
}
\arguments{
\item{query}{A TileDB Query object}

\item{attr}{A character value containing the attribute}
}
\value{
A integer with the number of elements in the results buffer
for the given attribute
}
\description{
Get TileDB Query result buffer elements
}
