% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ArraySchema.R
\name{tiledb_schema_object}
\alias{tiledb_schema_object}
\title{Succinctly describe a TileDB array schema}
\usage{
tiledb_schema_object(array)
}
\arguments{
\item{array}{A TileDB Array object}
}
\value{
A list containing two data frames, one describing the overall array as well as one
with descriptions about dimensions and attributes in the schema
}
\description{
This is an internal function that is not exported.
}
