% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{tiledb_sparse}
\alias{tiledb_sparse}
\title{Constructs a tiledb_sparse object backed by a persisted tiledb array uri}
\usage{
tiledb_sparse(
  uri,
  query_type = c("READ", "WRITE"),
  as.data.frame = FALSE,
  attrs = character(),
  extended = TRUE,
  ctx = tiledb_get_context()
)
}
\arguments{
\item{uri}{uri path to the tiledb dense array}

\item{query_type}{optionally loads the array in "READ" or "WRITE" only modes.}

\item{as.data.frame}{optional logical switch, defaults to "FALSE"}

\item{attrs}{optional character vector to select attributes, default is
empty implying all are selected}

\item{extended}{optional logical switch selecting wide \sQuote{data.frame}
format, defaults to "TRUE"}

\item{ctx}{tiledb_ctx (optional)}
}
\value{
tiledb_sparse array object
}
\description{
tiledb_sparse returns a list of coordinates and attributes vectors for reads
}
\section{Planned Deprecation}{

We plan to deprecate the \code{tiledb_sparse} array type in a future release. While exact
timelines have not been finalised, it is advised to the \code{tiledb_array} for both
\emph{dense} and \emph{sparse} arrays going forward.
}

