% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{tiledb_stats_raw_dump}
\alias{tiledb_stats_raw_dump}
\title{Dumps internal TileDB statistics as JSON to file}
\usage{
tiledb_stats_raw_dump(path)
}
\arguments{
\item{path}{Character variable with path to stats file;
if the empty string is passed then the result is displayed on stdout.}
}
\description{
Dumps internal TileDB statistics as JSON to file
}
\examples{
\dontshow{ctx <- tiledb_ctx(limitTileDBCores())}
pth <- tempfile()
tiledb_stats_raw_dump(pth)
cat(readLines(pth)[1:10], sep = "\n")

}
