% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SparseArray.R
\name{tiledb_subarray}
\alias{tiledb_subarray}
\title{Query a array using a subarray vector}
\usage{
tiledb_subarray(A, subarray_vector, attrs = c())
}
\arguments{
\item{A}{tiledb_sparse or tiledb_dense}

\item{subarray_vector}{subarray to query}

\item{attrs}{list of attributes to query}
}
\value{
list of attributes being returned with query results
}
\description{
tiledb_subarray returns a results of query
}
