% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subarray.R
\name{tiledb_subarray}
\alias{tiledb_subarray}
\title{Constructs a \code{tiledb_subarray} object from a TileDB Query}
\usage{
tiledb_subarray(query)
}
\arguments{
\item{query}{A TileDB Query Object}
}
\value{
tiledb_subarray object
}
\description{
Constructs a \code{tiledb_subarray} object from a TileDB Query
}
