% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Version.R
\name{tiledb_version}
\alias{tiledb_version}
\title{The version of the libtiledb library}
\usage{
tiledb_version()
}
\value{
An named int vector c(major, minor, patch)
}
\description{
The version of the libtiledb library
}
\examples{
tiledb_version()
}
