% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_create_dir}
\alias{tiledb_vfs_create_dir}
\title{Create a VFS Directory}
\usage{
tiledb_vfs_create_dir(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a diretory path}
}
\value{
The uri value of the created directory
}
\description{
Create a VFS Directory
}
