% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_file_size}
\alias{tiledb_vfs_file_size}
\title{Return VFS File Size}
\usage{
tiledb_vfs_file_size(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a file path}
}
\value{
The size removed file
}
\description{
Return VFS File Size
}
