% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_move_dir}
\alias{tiledb_vfs_move_dir}
\title{Move (or rename) a VFS Directory}
\usage{
tiledb_vfs_move_dir(olduri, newuri, vfs = tiledb_get_vfs())
}
\arguments{
\item{olduri}{Character variable with an existing URI describing a directory path}

\item{newuri}{Character variable with a new desired URI directory path}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The newuri value of the moved directory
}
\description{
Move (or rename) a VFS Directory
}
