% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_remove_bucket}
\alias{tiledb_vfs_remove_bucket}
\title{Remove a VFS Bucket}
\usage{
tiledb_vfs_remove_bucket(vfs, uri)
}
\arguments{
\item{vfs}{TileDB VFS object}

\item{uri}{Character variable with a URI describing a cloud bucket}
}
\value{
The uri value
}
\description{
Remove a VFS Bucket
}
