% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VFS.R
\name{tiledb_vfs_touch}
\alias{tiledb_vfs_touch}
\title{Touch a VFS URI Resource}
\usage{
tiledb_vfs_touch(uri, vfs = tiledb_get_vfs())
}
\arguments{
\item{uri}{Character variable with a URI describing a bucket, file or directory}

\item{vfs}{A TileDB VFS object; default is to use a cached value.}
}
\value{
The uri value
}
\description{
Touch a VFS URI Resource
}
